DifferentiationInterface

Documentation for DifferentiationInterface.

This is an interface to various autodiff backends for differentiating functions of the form f(x) = y, where x and y are either numbers or arrays.

Public

DifferentiationInterface.DifferentiationInterface — Module

DifferentiationInterface

An experimental redesign for AbstractDifferentiation.jl.

Exports

ChainRulesForwardBackend

ChainRulesReverseBackend

EnzymeForwardBackend

EnzymeReverseBackend

FiniteDiffBackend

ForwardDiffBackend

ReverseDiffBackend

pullback!

pushforward!

DifferentiationInterface.ChainRulesForwardBackend — Type

ChainRulesForwardBackend{RC}

Performs autodiff with forward-mode AD packages based on ChainRulesCore.jl, like Diffractor.jl.

This must be constructed with an appropriate RuleConfig instance.

using Diffractor, DifferentiationInterface
backend = ChainRulesForwardBackend(Diffractor.DiffractorRuleConfig())

DifferentiationInterface.ChainRulesReverseBackend — Type

ChainRulesReverseBackend{RC}

Performs autodiff with reverse-mode AD packages based on ChainRulesCore.jl, like Zygote.jl.

This must be constructed with an appropriate RuleConfig instance:

using Zygote, DifferentiationInterface
backend = ChainRulesReverseBackend(Zygote.ZygoteRuleConfig())

DifferentiationInterface.EnzymeForwardBackend — Type

EnzymeForwardBackend

Performs forward-mode autodiff with Enzyme.jl.

DifferentiationInterface.EnzymeReverseBackend — Type

EnzymeReverseBackend

Performs reverse-mode autodiff with Enzyme.jl.

DifferentiationInterface.FiniteDiffBackend — Type

FiniteDiffBackend

Performs autodiff with FiniteDiff.jl.

DifferentiationInterface.ForwardDiffBackend — Type

ForwardDiffBackend

Performs autodiff with ForwardDiff.jl.

DifferentiationInterface.ReverseDiffBackend — Type

ReverseDiffBackend

Performs autodiff with ReverseDiff.jl.

DifferentiationInterface.pullback! — Function

pullback!(dx, backend, f, x, dy[, stuff])

Compute a vector-Jacobian product inside dx and return it.

Arguments

dx: tangent, might be modified

backend: reverse-mode autodiff backend

f: function x -> y to differentiate

x: argument

dy: cotangent

stuff: optional backend-specific storage (cache, config), might be modified

DifferentiationInterface.pushforward! — Function

pushforward!(dy, backend, f, x, dx[, stuff])

Compute a Jacobian-vector product inside dy and return it.

Arguments

dy: cotangent, might be modified

backend: forward-mode autodiff backend

f: function x -> y to differentiate

x: argument

dx: tangent

stuff: optional backend-specific storage (cache, config), might be modified

Internals

Math

Some implementation reminders:

pushforward	pullback

scalar -> scalar	derivative multiplied by the tangent	derivative multiplied by the cotangent

scalar -> vector	derivative vector multiplied componentwise by the tangent vector	dot product between the derivative vector and the cotangent vector

vector -> scalar	dot product between the gradient vector and the tangent vector	gradient vector multiplied componentwise by the cotangent

vector -> vector	Jacobian matrix multiplied by the tangent vector	transposed Jacobian matrix multiplied by the cotangent vector

Index

DifferentiationInterface.DifferentiationInterface

DifferentiationInterface.ChainRulesForwardBackend

DifferentiationInterface.ChainRulesReverseBackend

DifferentiationInterface.EnzymeForwardBackend

DifferentiationInterface.EnzymeReverseBackend

DifferentiationInterface.FiniteDiffBackend

DifferentiationInterface.ForwardDiffBackend

DifferentiationInterface.ReverseDiffBackend

DifferentiationInterface.pullback!

DifferentiationInterface.pushforward!