Core API

DifferentiationInterface — Module

DifferentiationInterface

An interface to various automatic differentiation backends in Julia.

Exports

AutoChainRules

AutoDiffractor

AutoEnzyme

AutoFastDifferentiation

AutoFiniteDiff

AutoFiniteDifferences

AutoForwardDiff

AutoPolyesterForwardDiff

AutoReverseDiff

AutoSparseFastDifferentiation

AutoSparseFiniteDiff

AutoSparseForwardDiff

AutoSparsePolyesterForwardDiff

AutoSparseReverseDiff

AutoSparseZygote

AutoTapir

AutoTracker

AutoZygote

SecondOrder

check_available

check_hessian

check_mutation

derivative

derivative!!

gradient

gradient!!

hessian

hessian!!

hvp

hvp!!

jacobian

jacobian!!

prepare_derivative

prepare_gradient

prepare_hessian

prepare_hvp

prepare_jacobian

prepare_pullback

prepare_pushforward

prepare_second_derivative

pullback

pullback!!

pushforward

pushforward!!

second_derivative

second_derivative!!

value_and_derivative

value_and_derivative!!

value_and_gradient

value_and_gradient!!

value_and_jacobian

value_and_jacobian!!

value_and_pullback

value_and_pullback!!

value_and_pushforward

value_and_pushforward!!

Derivative

DifferentiationInterface.derivative — Function

derivative(f, backend, x, [extras]) -> der

DifferentiationInterface.derivative!! — Function

derivative!!(f, der, backend, x, [extras]) -> der

DifferentiationInterface.prepare_derivative — Method

prepare_derivative([other_extras], f, backend, x) -> extras
prepare_derivative([other_extras], f!, backend, y, x) -> extras

Create an extras object subtyping DerivativeExtras that can be given to derivative operators.

DifferentiationInterface.value_and_derivative — Function

value_and_derivative(f, backend, x, [extras]) -> (y, der)

DifferentiationInterface.value_and_derivative!! — Function

value_and_derivative!!(f!, y, der, backend, x, [extras]) -> (y, der)

DifferentiationInterface.value_and_derivative!! — Function

value_and_derivative!!(f, der, backend, x, [extras]) -> (y, der)

Gradient

DifferentiationInterface.gradient — Function

gradient(f, backend, x, [extras]) -> grad

DifferentiationInterface.gradient!! — Function

gradient!!(f, grad, backend, x, [extras]) -> grad

DifferentiationInterface.prepare_gradient — Method

prepare_gradient([other_extras], f, backend, x) -> extras

Create an extras object subtyping GradientExtras that can be given to gradient operators.

DifferentiationInterface.value_and_gradient — Function

value_and_gradient(f, backend, x, [extras]) -> (y, grad)

DifferentiationInterface.value_and_gradient!! — Function

value_and_gradient!!(f, grad, backend, x, [extras]) -> (y, grad)

Jacobian

DifferentiationInterface.jacobian — Function

jacobian(f, backend, x, [extras]) -> jac

DifferentiationInterface.jacobian!! — Function

jacobian!!(f, jac, backend, x, [extras]) -> jac

DifferentiationInterface.prepare_jacobian — Method

prepare_jacobian([other_extras], f, backend, x) -> extras
prepare_jacobian([other_extras], f!, backend, y, x) -> extras

Create an extras object subtyping JacobianExtras that can be given to Jacobian operators.

DifferentiationInterface.value_and_jacobian — Function

value_and_jacobian(f, backend, x, [extras]) -> (y, jac)

DifferentiationInterface.value_and_jacobian!! — Function

value_and_jacobian!!(f, jac, backend, x, [extras]) -> (y, jac)

DifferentiationInterface.value_and_jacobian!! — Function

value_and_jacobian!!(f!, y, jac, backend, x, [extras]) -> (y, jac)

Second order

DifferentiationInterface.SecondOrder — Type

SecondOrder

Combination of two backends for second-order differentiation.

Constructor

SecondOrder(outer, inner)

Fields

outer::ADTypes.AbstractADType: backend for the outer differentiation

inner::ADTypes.AbstractADType: backend for the inner differentiation

DifferentiationInterface.prepare_second_derivative — Method

prepare_second_derivative([other_extras], f, backend, x) -> extras

Create an extras object subtyping SecondDerivativeExtras that can be given to second derivative operators.

DifferentiationInterface.second_derivative — Function

second_derivative(f, backend, x, [extras]) -> der2

DifferentiationInterface.second_derivative!! — Function

second_derivative!!(f, der2, backend, x, [extras]) -> der2

DifferentiationInterface.hessian — Function

hessian(f, backend, x, [extras]) -> hess

DifferentiationInterface.hessian!! — Function

hessian!!(f, hess, backend, x, [extras]) -> hess

DifferentiationInterface.prepare_hessian — Method

prepare_hessian([other_extras], f, backend, x) -> extras

Create an extras object subtyping HessianExtras that can be given to Hessian operators.

DifferentiationInterface.hvp — Function

hvp(f, backend, x, v, [extras]) -> p

DifferentiationInterface.hvp!! — Function

hvp!!(f, p, backend, x, v, [extras]) -> p

DifferentiationInterface.prepare_hvp — Method

prepare_hvp([other_extras], f, backend, x) -> extras

Create an extras object subtyping HVPExtras that can be given to Hessian-vector product operators.

Primitives

DifferentiationInterface.prepare_pushforward — Method

prepare_pushforward([other_extras], f, backend, x) -> extras
prepare_pushforward([other_extras], f!, backend, y, x) -> extras

Create an extras object subtyping PushforwardExtras that can be given to pushforward operators.

DifferentiationInterface.pushforward — Function

pushforward(f, backend, x, dx, [extras]) -> (y, dy)

DifferentiationInterface.pushforward!! — Function

pushforward!!(f, dy, backend, x, dx, [extras]) -> (y, dy)

DifferentiationInterface.value_and_pushforward — Function

value_and_pushforward(f, backend, x, dx, [extras]) -> (y, dy)

Info

Required primitive for forward mode backends to support allocating functions.

DifferentiationInterface.value_and_pushforward!! — Function

value_and_pushforward!!(f!, y, dy, backend, x, dx, [extras]) -> (y, dy)

Info

Required primitive for forward mode backends to support mutating functions.

DifferentiationInterface.value_and_pushforward!! — Function

value_and_pushforward!!(f, dy, backend, x, dx, [extras]) -> (y, dy)

DifferentiationInterface.prepare_pullback — Method

prepare_pullback([other_extras], f, backend, x) -> extras
prepare_pullback([other_extras], f!, backend, y, x) -> extras

Create an extras object subtyping PullbackExtras that can be given to pullback operators.

DifferentiationInterface.pullback — Function

pullback(f, backend, x, dy, [extras]) -> dx

DifferentiationInterface.pullback!! — Function

pullback!!(f, dx, backend, x, dy, [extras]) -> dx

DifferentiationInterface.value_and_pullback — Function

value_and_pullback(f, backend, x, dy, [extras]) -> (y, dx)

Info

Required primitive for reverse mode backends to support allocating functions.

DifferentiationInterface.value_and_pullback!! — Function

value_and_pullback!!(f, dx, backend, x, dy, [extras]) -> (y, dx)

DifferentiationInterface.value_and_pullback!! — Function

value_and_pullback!!(f!, y, dx, backend, x, dy, [extras]) -> (y, dx)

Info

Required primitive for reverse mode backends to support mutating functions.

Backend queries

DifferentiationInterface.check_available — Method

check_available(backend)

Check whether backend is available by trying a scalar-to-scalar derivative.

Warning

Might take a while due to compilation time.

DifferentiationInterface.check_hessian — Method

check_hessian(backend)

Check whether backend supports second order differentiation by trying a hessian.

Warning

Might take a while due to compilation time.

DifferentiationInterface.check_mutation — Method

check_mutation(backend)

Check whether backend supports differentiation of mutating functions by trying a jacobian.

Warning

Might take a while due to compilation time.

Internals

This is not part of the public API.

DifferentiationInterface.DerivativeExtras — Type

DerivativeExtras

Abstract type for additional information needed by derivative operators.

DifferentiationInterface.ForwardOverForward — Type

ForwardOverForward

Traits identifying second-order backends that compute HVPs in forward over forward mode (inefficient).

DifferentiationInterface.ForwardOverReverse — Type

ForwardOverReverse

Traits identifying second-order backends that compute HVPs in forward over reverse mode.

DifferentiationInterface.GradientExtras — Type

GradientExtras

Abstract type for additional information needed by gradient operators.

DifferentiationInterface.HVPExtras — Type

HVPExtras

Abstract type for additional information needed by Hessian-vector product operators.

DifferentiationInterface.HessianExtras — Type

HessianExtras

Abstract type for additional information needed by Hessian operators.

DifferentiationInterface.JacobianExtras — Type

JacobianExtras

Abstract type for additional information needed by Jacobian operators.

DifferentiationInterface.MutationNotSupported — Type

MutationNotSupported

Trait identifying backends that do not support mutating functions f!(y, x).

DifferentiationInterface.MutationSupported — Type

MutationSupported

Trait identifying backends that support mutating functions f!(y, x).

DifferentiationInterface.PullbackExtras — Type

PullbackExtras

Abstract type for additional information needed by pullback operators.

DifferentiationInterface.PullbackFast — Type

PullbackFast

Trait identifying backends that support efficient pullbacks.

DifferentiationInterface.PullbackSlow — Type

PullbackSlow

Trait identifying backends that do not support efficient pullbacks.

DifferentiationInterface.PushforwardExtras — Type

PushforwardExtras

Abstract type for additional information needed by pushforward operators.

DifferentiationInterface.PushforwardFast — Type

PushforwardFast

Trait identifying backends that support efficient pushforwards.

DifferentiationInterface.PushforwardSlow — Type

PushforwardSlow

Trait identifying backends that do not support efficient pushforwards.

DifferentiationInterface.ReverseOverForward — Type

ReverseOverForward

Traits identifying second-order backends that compute HVPs in reverse over forward mode.

DifferentiationInterface.ReverseOverReverse — Type

ReverseOverReverse

Traits identifying second-order backends that compute HVPs in reverse over reverse mode.

DifferentiationInterface.SecondDerivativeExtras — Type

SecondDerivativeExtras

Abstract type for additional information needed by second derivative operators.

DifferentiationInterface.basis — Method

basis(backend, a::AbstractArray, i::CartesianIndex)

Construct the i-th stardard basis array in the vector space of a with element type eltype(a).

Note

If an AD backend benefits from a more specialized basis array implementation, this function can be extended on the backend type.

DifferentiationInterface.mode — Method

mode(backend)

Return the AD mode of a backend in a statically predictable way.

The return value is a Type object chosen among:

ADTypes.AbstractForwardMode

ADTypes.AbstractFiniteDifferencesMode

ADTypes.AbstractReverseMode

ADTypes.AbstractSymbolicDifferentiationMode

This function exists because there are backends (like Enzyme) that can support both forward and reverse mode, which means their ADTypes.jl object does not subtype either.

DifferentiationInterface.pullback_performance — Method

pullback_performance(backend)

Return PullbackFast or PullbackSlow in a statically predictable way.

DifferentiationInterface.pushforward_performance — Method

pushforward_performance(backend)

Return PushforwardFast or PushforwardSlow in a statically predictable way.

DifferentiationInterface.supports_mutation — Method

supports_mutation(backend)

Return MutationSupported or MutationNotSupported in a statically predictable way.