API

DifferentiationInterface — Module

DifferentiationInterface

An interface to various automatic differentiation backends in Julia.

Argument wrappers

DifferentiationInterface.Context — Type

Context

Abstract supertype for additional context arguments, which can be passed to differentiation operators after the active input x but are not differentiated.

Subtypes

Constant

Cache

ConstantOrCache

DifferentiationInterface.Constant — Type

Constant

Concrete type of Context argument which is kept constant during differentiation.

Note that an operator can be prepared with an arbitrary value of the constant. However, same-point preparation must occur with the exact value that will be reused later.

Warning

Some backends require any Constant context to be a Number or an AbstractArray.

Example

julia> using DifferentiationInterface

julia> import ForwardDiff

julia> f(x, c) = c * sum(abs2, x);

julia> gradient(f, AutoForwardDiff(), [1.0, 2.0], Constant(10))
2-element Vector{Float64}:
 20.0
 40.0

julia> gradient(f, AutoForwardDiff(), [1.0, 2.0], Constant(100))
2-element Vector{Float64}:
 200.0
 400.0

DifferentiationInterface.Cache — Type

Cache

Concrete type of Context argument which can be mutated with active values during differentiation.

The initial values present inside the cache do not matter.

For some backends, preparation allocates the required memory for Cache contexts with the right element type, similar to PreallocationTools.jl.

Warning

Some backends require any Cache context to be an AbstractArray, others accept nested (named) tuples of AbstractArrays.

Example

julia> using DifferentiationInterface

julia> import ForwardDiff

julia> f(x, c) = sum(copyto!(c, x));

julia> prep = prepare_gradient(f, AutoForwardDiff(), [1.0, 2.0], Cache(zeros(2)));

julia> gradient(f, prep, AutoForwardDiff(), [3.0, 4.0], Cache(zeros(2)))
2-element Vector{Float64}:
 1.0
 1.0

DifferentiationInterface.ConstantOrCache — Type

ConstantOrCache

Concrete type of Context argument which can contain a mixture of constants and caches, passed along to the backend without modification.

Unlike for Cache, it is up to the user to ensure that the internal storage can adapt to the required element types, for instance by using PreallocationTools.jl directly.

First order

Pushforward

DifferentiationInterface.prepare_pushforward — Function

prepare_pushforward(f,     backend, x, tx, [contexts...]; strict=Val(false)) -> prep
prepare_pushforward(f!, y, backend, x, tx, [contexts...]; strict=Val(false)) -> prep

Create a prep object that can be given to pushforward and its variants to speed them up.

Depending on the backend, this can have several effects (preallocating memory, recording an execution trace) which are transparent to the user.

For in-place functions, y is mutated by f! during preparation.

Warning

The preparation result prep is only reusable as long as the arguments to pushforward do not change type or size, and the function and backend themselves are not modified. Otherwise, preparation becomes invalid and you need to run it again. In some settings, invalid preparations may still give correct results (e.g. for backends that require no preparation), but this is not a semantic guarantee and should not be relied upon.

When strict=Val(true), type checking is enforced between preparation and execution (but size checking is left to the user).

DifferentiationInterface.prepare_pushforward_same_point — Function

prepare_pushforward_same_point(f,     backend, x, tx, [contexts...]; strict=Val(false)) -> prep_same
prepare_pushforward_same_point(f!, y, backend, x, tx, [contexts...]; strict=Val(false)) -> prep_same

Create a prep object that can be given to pushforward and its variants to speed them up, if they are applied at the same point x and with the same contexts.

Depending on the backend, this can have several effects (preallocating memory, recording an execution trace) which are transparent to the user.

For in-place functions, y is mutated by f! during preparation.

Warning

The preparation result prep is only reusable as long as the arguments to pushforward do not change type or size, and the function and backend themselves are not modified. Otherwise, preparation becomes invalid and you need to run it again. In some settings, invalid preparations may still give correct results (e.g. for backends that require no preparation), but this is not a semantic guarantee and should not be relied upon.

When strict=Val(true), type checking is enforced between preparation and execution (but size checking is left to the user).

DifferentiationInterface.pushforward — Function

pushforward(f,     [prep,] backend, x, tx, [contexts...]) -> ty
pushforward(f!, y, [prep,] backend, x, tx, [contexts...]) -> ty

Compute the pushforward of the function f at point x with a tuple of tangents tx.

To improve performance via operator preparation, refer to prepare_pushforward and prepare_pushforward_same_point.

Tip

Pushforwards are also commonly called Jacobian-vector products or JVPs. This function could have been named jvp.

DifferentiationInterface.pushforward! — Function

pushforward!(f,     dy, [prep,] backend, x, tx, [contexts...]) -> ty
pushforward!(f!, y, dy, [prep,] backend, x, tx, [contexts...]) -> ty

Compute the pushforward of the function f at point x with a tuple of tangents tx, overwriting ty.

To improve performance via operator preparation, refer to prepare_pushforward and prepare_pushforward_same_point.

Tip

Pushforwards are also commonly called Jacobian-vector products or JVPs. This function could have been named jvp!.

DifferentiationInterface.value_and_pushforward — Function

value_and_pushforward(f,     [prep,] backend, x, tx, [contexts...]) -> (y, ty)
value_and_pushforward(f!, y, [prep,] backend, x, tx, [contexts...]) -> (y, ty)

Compute the value and the pushforward of the function f at point x with a tuple of tangents tx.

To improve performance via operator preparation, refer to prepare_pushforward and prepare_pushforward_same_point.

Tip

Pushforwards are also commonly called Jacobian-vector products or JVPs. This function could have been named value_and_jvp.

Info

Required primitive for forward mode backends.

DifferentiationInterface.value_and_pushforward! — Function

value_and_pushforward!(f,     dy, [prep,] backend, x, tx, [contexts...]) -> (y, ty)
value_and_pushforward!(f!, y, dy, [prep,] backend, x, tx, [contexts...]) -> (y, ty)

Compute the value and the pushforward of the function f at point x with a tuple of tangents tx, overwriting ty.

To improve performance via operator preparation, refer to prepare_pushforward and prepare_pushforward_same_point.

Tip

Pushforwards are also commonly called Jacobian-vector products or JVPs. This function could have been named value_and_jvp!.

Pullback

DifferentiationInterface.prepare_pullback — Function

prepare_pullback(f,     backend, x, ty, [contexts...]; strict=Val(false)) -> prep
prepare_pullback(f!, y, backend, x, ty, [contexts...]; strict=Val(false)) -> prep

Create a prep object that can be given to pullback and its variants to speed them up.

Depending on the backend, this can have several effects (preallocating memory, recording an execution trace) which are transparent to the user.

For in-place functions, y is mutated by f! during preparation.

Warning

The preparation result prep is only reusable as long as the arguments to pullback do not change type or size, and the function and backend themselves are not modified. Otherwise, preparation becomes invalid and you need to run it again. In some settings, invalid preparations may still give correct results (e.g. for backends that require no preparation), but this is not a semantic guarantee and should not be relied upon.

When strict=Val(true), type checking is enforced between preparation and execution (but size checking is left to the user).

DifferentiationInterface.prepare_pullback_same_point — Function

prepare_pullback_same_point(f,     backend, x, ty, [contexts...]; strict=Val(false)) -> prep_same
prepare_pullback_same_point(f!, y, backend, x, ty, [contexts...]; strict=Val(false)) -> prep_same

Create a prep object that can be given to pullback and its variants to speed them up, if they are applied at the same point x and with the same contexts.

Depending on the backend, this can have several effects (preallocating memory, recording an execution trace) which are transparent to the user.

For in-place functions, y is mutated by f! during preparation.

Warning

The preparation result prep is only reusable as long as the arguments to pullback do not change type or size, and the function and backend themselves are not modified. Otherwise, preparation becomes invalid and you need to run it again. In some settings, invalid preparations may still give correct results (e.g. for backends that require no preparation), but this is not a semantic guarantee and should not be relied upon.

When strict=Val(true), type checking is enforced between preparation and execution (but size checking is left to the user).

DifferentiationInterface.pullback — Function

pullback(f,     [prep,] backend, x, ty, [contexts...]) -> tx
pullback(f!, y, [prep,] backend, x, ty, [contexts...]) -> tx

Compute the pullback of the function f at point x with a tuple of tangents ty.

To improve performance via operator preparation, refer to prepare_pullback and prepare_pullback_same_point.

Tip

Pullbacks are also commonly called vector-Jacobian products or VJPs. This function could have been named vjp.

DifferentiationInterface.pullback! — Function

pullback!(f,     dx, [prep,] backend, x, ty, [contexts...]) -> tx
pullback!(f!, y, dx, [prep,] backend, x, ty, [contexts...]) -> tx

Compute the pullback of the function f at point x with a tuple of tangents ty, overwriting dx.

To improve performance via operator preparation, refer to prepare_pullback and prepare_pullback_same_point.

Tip

Pullbacks are also commonly called vector-Jacobian products or VJPs. This function could have been named vjp!.

DifferentiationInterface.value_and_pullback — Function

value_and_pullback(f,     [prep,] backend, x, ty, [contexts...]) -> (y, tx)
value_and_pullback(f!, y, [prep,] backend, x, ty, [contexts...]) -> (y, tx)

Compute the value and the pullback of the function f at point x with a tuple of tangents ty.

To improve performance via operator preparation, refer to prepare_pullback and prepare_pullback_same_point.

Tip

Pullbacks are also commonly called vector-Jacobian products or VJPs. This function could have been named value_and_vjp.

Info

Required primitive for reverse mode backends.

DifferentiationInterface.value_and_pullback! — Function

value_and_pullback!(f,     dx, [prep,] backend, x, ty, [contexts...]) -> (y, tx)
value_and_pullback!(f!, y, dx, [prep,] backend, x, ty, [contexts...]) -> (y, tx)

Compute the value and the pullback of the function f at point x with a tuple of tangents ty, overwriting dx.

To improve performance via operator preparation, refer to prepare_pullback and prepare_pullback_same_point.

Tip

Pullbacks are also commonly called vector-Jacobian products or VJPs. This function could have been named value_and_vjp!.

Derivative

DifferentiationInterface.prepare_derivative — Function

prepare_derivative(f,     backend, x, [contexts...]; strict=Val(false)) -> prep
prepare_derivative(f!, y, backend, x, [contexts...]; strict=Val(false)) -> prep

Create a prep object that can be given to derivative and its variants to speed them up.

Depending on the backend, this can have several effects (preallocating memory, recording an execution trace) which are transparent to the user.

For in-place functions, y is mutated by f! during preparation.

Warning

The preparation result prep is only reusable as long as the arguments to derivative do not change type or size, and the function and backend themselves are not modified. Otherwise, preparation becomes invalid and you need to run it again. In some settings, invalid preparations may still give correct results (e.g. for backends that require no preparation), but this is not a semantic guarantee and should not be relied upon.

When strict=Val(true), type checking is enforced between preparation and execution (but size checking is left to the user).

DifferentiationInterface.derivative — Function

derivative(f,     [prep,] backend, x, [contexts...]) -> der
derivative(f!, y, [prep,] backend, x, [contexts...]) -> der

Compute the derivative of the function f at point x.

To improve performance via operator preparation, refer to prepare_derivative.

DifferentiationInterface.derivative! — Function

derivative!(f,     der, [prep,] backend, x, [contexts...]) -> der
derivative!(f!, y, der, [prep,] backend, x, [contexts...]) -> der

Compute the derivative of the function f at point x, overwriting der.

To improve performance via operator preparation, refer to prepare_derivative.

DifferentiationInterface.value_and_derivative — Function

value_and_derivative(f,     [prep,] backend, x, [contexts...]) -> (y, der)
value_and_derivative(f!, y, [prep,] backend, x, [contexts...]) -> (y, der)

Compute the value and the derivative of the function f at point x.

To improve performance via operator preparation, refer to prepare_derivative.

DifferentiationInterface.value_and_derivative! — Function

value_and_derivative!(f,     der, [prep,] backend, x, [contexts...]) -> (y, der)
value_and_derivative!(f!, y, der, [prep,] backend, x, [contexts...]) -> (y, der)

Compute the value and the derivative of the function f at point x, overwriting der.

To improve performance via operator preparation, refer to prepare_derivative.

Gradient

DifferentiationInterface.prepare_gradient — Function

prepare_gradient(f, backend, x, [contexts...]; strict=Val(false)) -> prep

Create a prep object that can be given to gradient and its variants to speed them up.

Depending on the backend, this can have several effects (preallocating memory, recording an execution trace) which are transparent to the user.

Warning

The preparation result prep is only reusable as long as the arguments to gradient do not change type or size, and the function and backend themselves are not modified. Otherwise, preparation becomes invalid and you need to run it again. In some settings, invalid preparations may still give correct results (e.g. for backends that require no preparation), but this is not a semantic guarantee and should not be relied upon.

When strict=Val(true), type checking is enforced between preparation and execution (but size checking is left to the user).

DifferentiationInterface.gradient — Function

gradient(f, [prep,] backend, x, [contexts...]) -> grad

Compute the gradient of the function f at point x.

To improve performance via operator preparation, refer to prepare_gradient.

DifferentiationInterface.gradient! — Function

gradient!(f, grad, [prep,] backend, x, [contexts...]) -> grad

Compute the gradient of the function f at point x, overwriting grad.

To improve performance via operator preparation, refer to prepare_gradient.

DifferentiationInterface.value_and_gradient — Function

value_and_gradient(f, [prep,] backend, x, [contexts...]) -> (y, grad)

Compute the value and the gradient of the function f at point x.

To improve performance via operator preparation, refer to prepare_gradient.

DifferentiationInterface.value_and_gradient! — Function

value_and_gradient!(f, grad, [prep,] backend, x, [contexts...]) -> (y, grad)

Compute the value and the gradient of the function f at point x, overwriting grad.

To improve performance via operator preparation, refer to prepare_gradient.

Jacobian

DifferentiationInterface.prepare_jacobian — Function

prepare_jacobian(f,     backend, x, [contexts...]; strict=Val(false)) -> prep
prepare_jacobian(f!, y, backend, x, [contexts...]; strict=Val(false)) -> prep

Create a prep object that can be given to jacobian and its variants to speed them up.

Depending on the backend, this can have several effects (preallocating memory, recording an execution trace) which are transparent to the user.

For in-place functions, y is mutated by f! during preparation.

Warning

The preparation result prep is only reusable as long as the arguments to jacobian do not change type or size, and the function and backend themselves are not modified. Otherwise, preparation becomes invalid and you need to run it again. In some settings, invalid preparations may still give correct results (e.g. for backends that require no preparation), but this is not a semantic guarantee and should not be relied upon.

When strict=Val(true), type checking is enforced between preparation and execution (but size checking is left to the user).

DifferentiationInterface.jacobian — Function

jacobian(f,     [prep,] backend, x, [contexts...]) -> jac
jacobian(f!, y, [prep,] backend, x, [contexts...]) -> jac

Compute the Jacobian matrix of the function f at point x.

To improve performance via operator preparation, refer to prepare_jacobian.

DifferentiationInterface.jacobian! — Function

jacobian!(f,     jac, [prep,] backend, x, [contexts...]) -> jac
jacobian!(f!, y, jac, [prep,] backend, x, [contexts...]) -> jac

Compute the Jacobian matrix of the function f at point x, overwriting jac.

To improve performance via operator preparation, refer to prepare_jacobian.

DifferentiationInterface.value_and_jacobian — Function

value_and_jacobian(f,     [prep,] backend, x, [contexts...]) -> (y, jac)
value_and_jacobian(f!, y, [prep,] backend, x, [contexts...]) -> (y, jac)

Compute the value and the Jacobian matrix of the function f at point x.

To improve performance via operator preparation, refer to prepare_jacobian.

DifferentiationInterface.value_and_jacobian! — Function

value_and_jacobian!(f,     jac, [prep,] backend, x, [contexts...]) -> (y, jac)
value_and_jacobian!(f!, y, jac, [prep,] backend, x, [contexts...]) -> (y, jac)

Compute the value and the Jacobian matrix of the function f at point x, overwriting jac.

To improve performance via operator preparation, refer to prepare_jacobian.

Second order

DifferentiationInterface.SecondOrder — Type

SecondOrder

Combination of two backends for second-order differentiation.

Danger

SecondOrder backends do not support first-order operators.

Constructor

SecondOrder(outer_backend, inner_backend)

Fields

outer::AbstractADType: backend for the outer differentiation

inner::AbstractADType: backend for the inner differentiation

Second derivative

DifferentiationInterface.prepare_second_derivative — Function

prepare_second_derivative(f, backend, x, [contexts...]; strict=Val(false)) -> prep

Create a prep object that can be given to second_derivative and its variants to speed them up.

Depending on the backend, this can have several effects (preallocating memory, recording an execution trace) which are transparent to the user.

Warning

The preparation result prep is only reusable as long as the arguments to second_derivative do not change type or size, and the function and backend themselves are not modified. Otherwise, preparation becomes invalid and you need to run it again. In some settings, invalid preparations may still give correct results (e.g. for backends that require no preparation), but this is not a semantic guarantee and should not be relied upon.

When strict=Val(true), type checking is enforced between preparation and execution (but size checking is left to the user).

DifferentiationInterface.second_derivative — Function

second_derivative(f, [prep,] backend, x, [contexts...]) -> der2

Compute the second derivative of the function f at point x.

To improve performance via operator preparation, refer to prepare_second_derivative.

DifferentiationInterface.second_derivative! — Function

second_derivative!(f, der2, [prep,] backend, x, [contexts...]) -> der2

Compute the second derivative of the function f at point x, overwriting der2.

To improve performance via operator preparation, refer to prepare_second_derivative.

DifferentiationInterface.value_derivative_and_second_derivative — Function

value_derivative_and_second_derivative(f, [prep,] backend, x, [contexts...]) -> (y, der, der2)

Compute the value, first derivative and second derivative of the function f at point x.

To improve performance via operator preparation, refer to prepare_second_derivative.

DifferentiationInterface.value_derivative_and_second_derivative! — Function

value_derivative_and_second_derivative!(f, der, der2, [prep,] backend, x, [contexts...]) -> (y, der, der2)

Compute the value, first derivative and second derivative of the function f at point x, overwriting der and der2.

To improve performance via operator preparation, refer to prepare_second_derivative.

Hessian-vector product

DifferentiationInterface.prepare_hvp — Function

prepare_hvp(f, backend, x, tx, [contexts...]; strict=Val(false)) -> prep

Create a prep object that can be given to hvp and its variants to speed them up.

Depending on the backend, this can have several effects (preallocating memory, recording an execution trace) which are transparent to the user.

Warning

The preparation result prep is only reusable as long as the arguments to hvp do not change type or size, and the function and backend themselves are not modified. Otherwise, preparation becomes invalid and you need to run it again. In some settings, invalid preparations may still give correct results (e.g. for backends that require no preparation), but this is not a semantic guarantee and should not be relied upon.

When strict=Val(true), type checking is enforced between preparation and execution (but size checking is left to the user).

DifferentiationInterface.prepare_hvp_same_point — Function

prepare_hvp_same_point(f, backend, x, tx, [contexts...]; strict=Val(false)) -> prep_same

Create a prep object that can be given to hvp and its variants to speed them up, if they are applied at the same point x and with the same contexts.

Depending on the backend, this can have several effects (preallocating memory, recording an execution trace) which are transparent to the user.

Warning

The preparation result prep is only reusable as long as the arguments to hvp do not change type or size, and the function and backend themselves are not modified. Otherwise, preparation becomes invalid and you need to run it again. In some settings, invalid preparations may still give correct results (e.g. for backends that require no preparation), but this is not a semantic guarantee and should not be relied upon.

When strict=Val(true), type checking is enforced between preparation and execution (but size checking is left to the user).

DifferentiationInterface.hvp — Function

hvp(f, [prep,] backend, x, tx, [contexts...]) -> tg

Compute the Hessian-vector product of f at point x with a tuple of tangents tx.

To improve performance via operator preparation, refer to prepare_hvp and prepare_hvp_same_point.

DifferentiationInterface.hvp! — Function

hvp!(f, tg, [prep,] backend, x, tx, [contexts...]) -> tg

Compute the Hessian-vector product of f at point x with a tuple of tangents tx, overwriting tg.

To improve performance via operator preparation, refer to prepare_hvp and prepare_hvp_same_point.

DifferentiationInterface.gradient_and_hvp — Function

gradient_and_hvp(f, [prep,] backend, x, tx, [contexts...]) -> (grad, tg)

Compute the gradient and the Hessian-vector product of f at point x with a tuple of tangents tx.

To improve performance via operator preparation, refer to prepare_hvp and prepare_hvp_same_point.

DifferentiationInterface.gradient_and_hvp! — Function

gradient_and_hvp!(f, grad, tg, [prep,] backend, x, tx, [contexts...]) -> (grad, tg)

Compute the gradient and the Hessian-vector product of f at point x with a tuple of tangents tx, overwriting grad and tg.

To improve performance via operator preparation, refer to prepare_hvp and prepare_hvp_same_point.

Hessian

DifferentiationInterface.prepare_hessian — Function

prepare_hessian(f, backend, x, [contexts...]; strict=Val(false)) -> prep

Create a prep object that can be given to hessian and its variants to speed them up.

Depending on the backend, this can have several effects (preallocating memory, recording an execution trace) which are transparent to the user.

Warning

The preparation result prep is only reusable as long as the arguments to hessian do not change type or size, and the function and backend themselves are not modified. Otherwise, preparation becomes invalid and you need to run it again. In some settings, invalid preparations may still give correct results (e.g. for backends that require no preparation), but this is not a semantic guarantee and should not be relied upon.

When strict=Val(true), type checking is enforced between preparation and execution (but size checking is left to the user).

DifferentiationInterface.hessian — Function

hessian(f, [prep,] backend, x, [contexts...]) -> hess

Compute the Hessian matrix of the function f at point x.

To improve performance via operator preparation, refer to prepare_hessian.

DifferentiationInterface.hessian! — Function

hessian!(f, hess, [prep,] backend, x, [contexts...]) -> hess

Compute the Hessian matrix of the function f at point x, overwriting hess.

To improve performance via operator preparation, refer to prepare_hessian.

DifferentiationInterface.value_gradient_and_hessian — Function

value_gradient_and_hessian(f, [prep,] backend, x, [contexts...]) -> (y, grad, hess)

Compute the value, gradient vector and Hessian matrix of the function f at point x.

To improve performance via operator preparation, refer to prepare_hessian.

DifferentiationInterface.value_gradient_and_hessian! — Function

value_gradient_and_hessian!(f, grad, hess, [prep,] backend, x, [contexts...]) -> (y, grad, hess)

Compute the value, gradient vector and Hessian matrix of the function f at point x, overwriting grad and hess.

To improve performance via operator preparation, refer to prepare_hessian.

Utilities

Backend queries

DifferentiationInterface.check_available — Function

check_available(backend)

Check whether backend is available (i.e. whether the extension is loaded) and return a Bool.

DifferentiationInterface.check_inplace — Function

check_inplace(backend)

Check whether backend supports differentiation of in-place functions and return a Bool.

DifferentiationInterface.outer — Function

outer(backend::SecondOrder)
outer(backend::AbstractADType)

Return the outer backend of a SecondOrder object, tasked with differentiation at the second order.

For any other backend type, this function acts like the identity.

DifferentiationInterface.inner — Function

inner(backend::SecondOrder)
inner(backend::AbstractADType)

Return the inner backend of a SecondOrder object, tasked with differentiation at the first order.

For any other backend type, this function acts like the identity.

Backend switch

DifferentiationInterface.DifferentiateWith — Type

DifferentiateWith

Function wrapper that enforces differentiation with a "substitute" AD backend, possible different from the "true" AD backend that is called.

For instance, suppose a function f is not differentiable with Zygote because it involves mutation, but you know that it is differentiable with Enzyme. Then f2 = DifferentiateWith(f, AutoEnzyme()) is a new function that behaves like f, except that f2 is differentiable with Zygote (thanks to a chain rule which calls Enzyme under the hood). Moreover, any larger algorithm alg that calls f2 instead of f will also be differentiable with Zygote (as long as f was the only Zygote blocker).

Tip

This is mainly relevant for package developers who want to produce differentiable code at low cost, without writing the differentiation rules themselves. If you sprinkle a few DifferentiateWith in places where some AD backends may struggle, end users can pick from a wider variety of packages to differentiate your algorithms.

Warning

DifferentiateWith only supports out-of-place functions y = f(x) without additional context arguments. It only makes these functions differentiable if the true backend is either ForwardDiff or automatically importing rules from ChainRules (e.g. Zygote). Some backends are also able to manually import rules from ChainRules. For any other true backend, the differentiation behavior is not altered by DifferentiateWith (it becomes a transparent wrapper).

Fields

f: the function in question, with signature f(x)

backend::AbstractADType: the substitute backend to use for differentiation

Note

For the substitute AD backend to be called under the hood, its package needs to be loaded in addition to the package of the true AD backend.

Constructor

DifferentiateWith(f, backend)

Example

julia> using DifferentiationInterface

julia> import FiniteDiff, ForwardDiff, Zygote

julia> function f(x::Vector{Float64})
           a = Vector{Float64}(undef, 1)  # type constraint breaks ForwardDiff
           a[1] = sum(abs2, x)  # mutation breaks Zygote
           return a[1]
       end;

julia> f2 = DifferentiateWith(f, AutoFiniteDiff());

julia> f([3.0, 5.0]) == f2([3.0, 5.0])
true

julia> alg(x) = 7 * f2(x);

julia> ForwardDiff.gradient(alg, [3.0, 5.0])
2-element Vector{Float64}:
 42.0
 70.0

julia> Zygote.gradient(alg, [3.0, 5.0])[1]
2-element Vector{Float64}:
 42.0
 70.0

Sparsity tools

DifferentiationInterface.MixedMode — Type

MixedMode

Combination of a forward and a reverse mode backend for mixed-mode sparse Jacobian computation.

Danger

MixedMode backends only support jacobian and its variants, and it should be used inside an AutoSparse wrapper.

Constructor

MixedMode(forward_backend, reverse_backend)

DifferentiationInterface.DenseSparsityDetector — Type

DenseSparsityDetector

Sparsity pattern detector satisfying the detection API of ADTypes.jl.

The nonzeros in a Jacobian or Hessian are detected by computing the relevant matrix with dense AD, and thresholding the entries with a given tolerance (which can be numerically inaccurate). This process can be very slow, and should only be used if its output can be exploited multiple times to compute many sparse matrices.

Danger

In general, the sparsity pattern you obtain can depend on the provided input x. If you want to reuse the pattern, make sure that it is input-agnostic.

Warning

DenseSparsityDetector functionality is now located in a package extension, please load the SparseArrays.jl standard library before you use it.

Fields

backend::AbstractADType is the dense AD backend used under the hood

atol::Float64 is the minimum magnitude of a matrix entry to be considered nonzero

Constructor

DenseSparsityDetector(backend; atol, method=:iterative)

The keyword argument method::Symbol can be either:

:iterative: compute the matrix in a sequence of matrix-vector products (memory-efficient)

:direct: compute the matrix all at once (memory-hungry but sometimes faster).

Note that the constructor is type-unstable because method ends up being a type parameter of the DenseSparsityDetector object (this is not part of the API and might change).

Examples

using ADTypes, DifferentiationInterface, SparseArrays
import ForwardDiff

detector = DenseSparsityDetector(AutoForwardDiff(); atol=1e-5, method=:direct)

ADTypes.jacobian_sparsity(diff, rand(5), detector)

# output

4×5 SparseMatrixCSC{Bool, Int64} with 8 stored entries:
 1  1  ⋅  ⋅  ⋅
 ⋅  1  1  ⋅  ⋅
 ⋅  ⋅  1  1  ⋅
 ⋅  ⋅  ⋅  1  1

Sometimes the sparsity pattern is input-dependent:

ADTypes.jacobian_sparsity(x -> [prod(x)], rand(2), detector)

# output

1×2 SparseMatrixCSC{Bool, Int64} with 2 stored entries:
 1  1

ADTypes.jacobian_sparsity(x -> [prod(x)], [0, 1], detector)

# output

1×2 SparseMatrixCSC{Bool, Int64} with 1 stored entry:
 1  ⋅

Internals

The following is not part of the public API.

DifferentiationInterface.AutoForwardFromPrimitive — Type

AutoForwardFromPrimitive

Wrapper which forces a given backend to act as a reverse-mode backend.

Used in internal testing.

DifferentiationInterface.AutoReverseFromPrimitive — Type

AutoReverseFromPrimitive

Wrapper which forces a given backend to act as a reverse-mode backend.

Used in internal testing.

DifferentiationInterface.AutoSimpleFiniteDiff — Type

AutoSimpleFiniteDiff <: ADTypes.AbstractADType

Forward mode backend based on the finite difference (f(x + ε) - f(x)) / ε, with artificial chunk size to mimick ForwardDiff.

Constructor

AutoSimpleFiniteDiff(ε=1e-5; chunksize=nothing)

DifferentiationInterface.AutoZeroForward — Type

AutoZeroForward <: ADTypes.AbstractADType

Trivial backend that sets all derivatives to zero. Used in testing and benchmarking.

DifferentiationInterface.AutoZeroReverse — Type

AutoZeroReverse <: ADTypes.AbstractADType

Trivial backend that sets all derivatives to zero. Used in testing and benchmarking.

DifferentiationInterface.BatchSizeSettings — Type

BatchSizeSettings{B,singlebatch,aligned}

Configuration for the batch size deduced from a backend and a sample array of length N.

Type parameters

B::Int: batch size

singlebatch::Bool: whether B == N (B > N is not allowed)

aligned::Bool: whether N % B == 0

Fields

N::Int: array length

A::Int: number of batches A = div(N, B, RoundUp)

B_last::Int: size of the last batch (if aligned is false)

DifferentiationInterface.DerivativePrep — Type

DerivativePrep

Abstract type for additional information needed by derivative and its variants.

DifferentiationInterface.DontPrepareInner — Type

DontPrepareInner

Trait identifying outer backends for which the inner backend in second-order autodiff should not be prepared at all.

DifferentiationInterface.FixTail — Type

FixTail

Closure around a function f and a set of tail argument tail_args such that

(ft::FixTail)(args...) = ft.f(args..., ft.tail_args...)

DifferentiationInterface.ForwardAndReverseMode — Type

ForwardAndReverseMode <: ADTypes.AbstractMode

Appropriate mode type for MixedMode backends.

DifferentiationInterface.ForwardOverForward — Type

ForwardOverForward

Traits identifying second-order backends that compute HVPs in forward over forward mode (inefficient).

DifferentiationInterface.ForwardOverReverse — Type

ForwardOverReverse

Traits identifying second-order backends that compute HVPs in forward over reverse mode.

DifferentiationInterface.FunctionContext — Type

FunctionContext

Private type of Context argument used for passing functions inside second-order differentiation.

Behaves differently for Enzyme only, where the function can be annotated.

DifferentiationInterface.GradientPrep — Type

GradientPrep

Abstract type for additional information needed by gradient and its variants.

DifferentiationInterface.HVPPrep — Type

HVPPrep

Abstract type for additional information needed by hvp and its variants.

DifferentiationInterface.HessianPrep — Type

HessianPrep

Abstract type for additional information needed by hessian and its variants.

DifferentiationInterface.InPlaceNotSupported — Type

InPlaceNotSupported

Trait identifying backends that do not support in-place functions f!(y, x).

DifferentiationInterface.InPlaceSupported — Type

InPlaceSupported

Trait identifying backends that support in-place functions f!(y, x).

DifferentiationInterface.JacobianPrep — Type

JacobianPrep

Abstract type for additional information needed by jacobian and its variants.

DifferentiationInterface.PrepareInnerOverload — Type

PrepareInnerOverload

Trait identifying outer backends for which the inner backend in second-order autodiff should be prepared with an overloaded input type.

DifferentiationInterface.PrepareInnerSimple — Type

PrepareInnerSimple

Trait identifying outer backends for which the inner backend in second-order autodiff should be prepared with the same input type.

DifferentiationInterface.PullbackFast — Type

PullbackFast

Trait identifying backends that support efficient pullbacks.

DifferentiationInterface.PullbackPrep — Type

PullbackPrep

Abstract type for additional information needed by pullback and its variants.

DifferentiationInterface.PullbackSlow — Type

PullbackSlow

Trait identifying backends that do not support efficient pullbacks.

DifferentiationInterface.PushforwardFast — Type

PushforwardFast

Trait identifying backends that support efficient pushforwards.

DifferentiationInterface.PushforwardPrep — Type

PushforwardPrep

Abstract type for additional information needed by pushforward and its variants.

DifferentiationInterface.PushforwardSlow — Type

PushforwardSlow

Trait identifying backends that do not support efficient pushforwards.

DifferentiationInterface.ReverseOverForward — Type

ReverseOverForward

Traits identifying second-order backends that compute HVPs in reverse over forward mode.

DifferentiationInterface.ReverseOverReverse — Type

ReverseOverReverse

Traits identifying second-order backends that compute HVPs in reverse over reverse mode.

DifferentiationInterface.Rewrap — Type

Rewrap

Utility for recording context types of additional arguments (e.g. Constant or Cache) and re-wrapping them into their types after they have been unwrapped.

Useful for second-order differentiation.

DifferentiationInterface.SecondDerivativePrep — Type

SecondDerivativePrep

Abstract type for additional information needed by second_derivative and its variants.

ADTypes.mode — Method

mode(backend::SecondOrder)

Return the outer mode of the second-order backend.

DifferentiationInterface.basis — Method

basis(a::AbstractArray, i)

Construct the i-th standard basis array in the vector space of a.

DifferentiationInterface.fix_tail — Method

fix_tail(f, tail_args...)

Convenience for constructing a FixTail, with a shortcut when there are no tail arguments.

DifferentiationInterface.forward_backend — Method

forward_backend(m::MixedMode)

Return the forward-mode part of a MixedMode backend.

DifferentiationInterface.get_pattern — Method

get_pattern(M::AbstractMatrix)

Return the Bool-valued sparsity pattern for a given matrix.

Only specialized on SparseMatrixCSC because it is used with symbolic backends, and at the moment their sparse Jacobian/Hessian utilities return a SparseMatrixCSC.

The trivial dense fallback is designed to protect against a change of format in these packages.

DifferentiationInterface.hessian_sparsity_with_contexts — Method

hessian_sparsity_with_contexts(f, detector, x, contexts...)

Wrapper around ADTypes.hessian_sparsity enabling the allocation of caches with proper element types.

DifferentiationInterface.hvp_mode — Method

hvp_mode(backend)

Return the best combination of modes for hvp and its variants, among the following options:

ForwardOverForward

ForwardOverReverse

ReverseOverForward

ReverseOverReverse

DifferentiationInterface.inner_preparation_behavior — Method

inner_preparation_behavior(backend::AbstractADType)

Return PrepareInnerSimple, PrepareInnerOverload or DontPrepareInner in a statically predictable way.

DifferentiationInterface.inplace_support — Method

inplace_support(backend)

Return InPlaceSupported or InPlaceNotSupported in a statically predictable way.

DifferentiationInterface.ismutable_array — Method

ismutable_array(x)

Check whether x is a mutable array and return a Bool.

At the moment, this only returns false for StaticArrays.SArray.

DifferentiationInterface.jacobian_sparsity_with_contexts — Method

jacobian_sparsity_with_contexts(f, detector, x, contexts...)
jacobian_sparsity_with_contexts(f!, y, detector, x, contexts...)

Wrapper around ADTypes.jacobian_sparsity enabling the allocation of caches with proper element types.

DifferentiationInterface.multibasis — Method

multibasis(a::AbstractArray, inds)

Construct the sum of the i-th standard basis arrays in the vector space of a for all i ∈ inds.

Warning

Does not work on GPU, since this is intended for sparse autodiff and SparseMatrixColorings.jl doesn't work on GPUs either.

DifferentiationInterface.overloaded_input_type — Function

overloaded_input_type(prep)

If it exists, return the overloaded input type which will be passed to the differentiated function when preparation result prep is reused.

Danger

This function is experimental and not part of the public API.

DifferentiationInterface.pick_batchsize — Method

pick_batchsize(backend, x_or_y::AbstractArray)

Return a BatchSizeSettings appropriate for arrays of the same length as x_or_y with a given backend.

Note that the array in question can be either the input or the output of the function, depending on whether the backend performs forward- or reverse-mode AD.

DifferentiationInterface.pick_batchsize — Method

pick_batchsize(backend, N::Integer)

Return a BatchSizeSettings appropriate for arrays of length N with a given backend.

DifferentiationInterface.prepare!_derivative — Method

prepare!_derivative(f,     prep, backend, x, [contexts...]) -> new_prep
prepare!_derivative(f!, y, prep, backend, x, [contexts...]) -> new_prep

Same behavior as prepare_derivative but can resize the contents of an existing prep object to avoid some allocations.

There is no guarantee that prep will be mutated, or that performance will be improved compared to preparation from scratch.

Danger

Compared to when prep was first created, the only authorized modification is a size change for input x or output y. Any other modification (like a change of type for the input) is not supported and will give erroneous results.

Danger

For efficiency, this function needs to rely on backend package internals, therefore it not protected by semantic versioning.

DifferentiationInterface.prepare!_gradient — Method

prepare!_gradient(f, prep, backend, x, [contexts...]) -> new_prep

Same behavior as prepare_gradient but can resize the contents of an existing prep object to avoid some allocations.

There is no guarantee that prep will be mutated, or that performance will be improved compared to preparation from scratch.

Danger

Compared to when prep was first created, the only authorized modification is a size change for input x or output y. Any other modification (like a change of type for the input) is not supported and will give erroneous results.

Danger

For efficiency, this function needs to rely on backend package internals, therefore it not protected by semantic versioning.

DifferentiationInterface.prepare!_hessian — Method

prepare!_hessian(f, backend, x, [contexts...]) -> new_prep

Same behavior as prepare_hessian but can resize the contents of an existing prep object to avoid some allocations.

There is no guarantee that prep will be mutated, or that performance will be improved compared to preparation from scratch.

Danger

Compared to when prep was first created, the only authorized modification is a size change for input x or output y. Any other modification (like a change of type for the input) is not supported and will give erroneous results.

Danger

For efficiency, this function needs to rely on backend package internals, therefore it not protected by semantic versioning.

DifferentiationInterface.prepare!_hvp — Method

prepare!_hvp(f, backend, x, tx, [contexts...]) -> new_prep

Create a prep object that can be given to hvp and its variants to speed them up.

Depending on the backend, this can have several effects (preallocating memory, recording an execution trace) which are transparent to the user.

Warning

The preparation result prep is only reusable as long as the arguments to hvp do not change type or size, and the function and backend themselves are not modified. Otherwise, preparation becomes invalid and you need to run it again. In some settings, invalid preparations may still give correct results (e.g. for backends that require no preparation), but this is not a semantic guarantee and should not be relied upon.

When strict=Val(true), type checking is enforced between preparation and execution (but size checking is left to the user).

DifferentiationInterface.prepare!_jacobian — Method

prepare!_jacobian(f,     prep, backend, x, [contexts...]) -> new_prep
prepare!_jacobian(f!, y, prep, backend, x, [contexts...]) -> new_prep

Same behavior as prepare_jacobian but can resize the contents of an existing prep object to avoid some allocations.

There is no guarantee that prep will be mutated, or that performance will be improved compared to preparation from scratch.

Danger

Compared to when prep was first created, the only authorized modification is a size change for input x or output y. Any other modification (like a change of type for the input) is not supported and will give erroneous results.

Danger

For efficiency, this function needs to rely on backend package internals, therefore it not protected by semantic versioning.

DifferentiationInterface.prepare!_pullback — Method

prepare!_pullback(f,     prep, backend, x, ty, [contexts...]) -> new_prep
prepare!_pullback(f!, y, prep, backend, x, ty, [contexts...]) -> new_prep

Same behavior as prepare_pullback but can resize the contents of an existing prep object to avoid some allocations.

There is no guarantee that prep will be mutated, or that performance will be improved compared to preparation from scratch.

Danger

Compared to when prep was first created, the only authorized modification is a size change for input x or output y. Any other modification (like a change of type for the input) is not supported and will give erroneous results.

Danger

For efficiency, this function needs to rely on backend package internals, therefore it not protected by semantic versioning.

DifferentiationInterface.prepare!_pushforward — Method

prepare!_pushforward(f,     prep, backend, x, tx, [contexts...]) -> new_prep
prepare!_pushforward(f!, y, prep, backend, x, tx, [contexts...]) -> new_prep

Same behavior as prepare_pushforward but can resize the contents of an existing prep object to avoid some allocations.

There is no guarantee that prep will be mutated, or that performance will be improved compared to preparation from scratch.

Danger

Compared to when prep was first created, the only authorized modification is a size change for input x or output y. Any other modification (like a change of type for the input) is not supported and will give erroneous results.

Danger

For efficiency, this function needs to rely on backend package internals, therefore it not protected by semantic versioning.

DifferentiationInterface.prepare!_second_derivative — Method

prepare!_second_derivative(f, prep, backend, x, [contexts...]) -> new_prep

Same behavior as prepare_second_derivative but can resize the contents of an existing prep object to avoid some allocations.

There is no guarantee that prep will be mutated, or that performance will be improved compared to preparation from scratch.

Danger

Compared to when prep was first created, the only authorized modification is a size change for input x or output y. Any other modification (like a change of type for the input) is not supported and will give erroneous results.

Danger

For efficiency, this function needs to rely on backend package internals, therefore it not protected by semantic versioning.

DifferentiationInterface.pullback_performance — Method

pullback_performance(backend)

Return PullbackFast or PullbackSlow in a statically predictable way.

DifferentiationInterface.pushforward_performance — Method

pushforward_performance(backend)

Return PushforwardFast or PushforwardSlow in a statically predictable way.

DifferentiationInterface.reasonable_batchsize — Method

reasonable_batchsize(N::Integer, Bmax::Integer)

Reproduces the heuristic from ForwardDiff to minimize

the number of batches necessary to cover an array of length N

the number of leftover indices in the last partial batch

Source: https://github.com/JuliaDiff/ForwardDiff.jl/blob/ec74fbc32b10bbf60b3c527d8961666310733728/src/prelude.jl#L19-L29

DifferentiationInterface.recursive_similar — Method

recursive_similar(x, T)

Apply similar(_, T) recursively to x or its components.

Works if x is an AbstractArray or a (nested) NTuple / NamedTuple of AbstractArrays.

DifferentiationInterface.reverse_backend — Method

reverse_backend(m::MixedMode)

Return the reverse-mode part of a MixedMode backend.

DifferentiationInterface.threshold_batchsize — Function

threshold_batchsize(backend::AbstractADType, B::Integer)

If the backend object has a fixed batch size B0, return a new backend where the fixed batch size is min(B0, B). Otherwise, act as the identity.