Backends

Types

The possible backend choices are defined by ADTypes.jl.

Warning

Only the backends listed here are supported by DifferentiationInterface.jl, even though ADTypes.jl defines more.

ADTypes.AutoChainRules — Type

AutoChainRules{RC}

Chooses any AD library based on ChainRulesCore.jl, given an appropriate RuleConfig object.

Fields

ruleconfig::RC

ADTypes.AutoDiffractor — Type

AutoDiffractor

Chooses Diffractor.jl.

ADTypes.AutoEnzyme — Type

AutoEnzyme(Val(:forward))
AutoEnzyme(Val(:reverse))

Construct a forward or reverse mode AutoEnzyme backend.

Warning

This is the mode convention chosen by DifferentiationInterface.jl, for lack of a global consensus (see ADTypes.jl#24).

AutoEnzyme{M}

Chooses Enzyme.jl.

Fields

mode::M = nothing

ADTypes.AutoForwardDiff — Type

AutoForwardDiff{chunksize,T}

Chooses ForwardDiff.jl.

Fields

tag::T

ADTypes.AutoForwardDiff — Method

AutoForwardDiff(; chunksize = nothing, tag = nothing)

Constructor.

ADTypes.AutoFiniteDiff — Type

AutoFiniteDiff{T1,T2,T3}

Chooses FiniteDiff.jl.

Fields

fdtype::T1 = Val(:forward)

fdjtype::T2 = fdtype

fdhtype::T3 = Val(:hcentral)

ADTypes.AutoPolyesterForwardDiff — Type

AutoPolyesterForwardDiff{chunksize}

Chooses PolyesterForwardDiff.jl.

ADTypes.AutoPolyesterForwardDiff — Method

AutoPolyesterForwardDiff(; chunksize = nothing)

Constructor.

ADTypes.AutoReverseDiff — Type

AutoReverseDiff

Chooses ReverseDiff.jl.

Fields

compile::Bool = false

ADTypes.AutoZygote — Type

AutoZygote

Chooses Zygote.jl.

Package extensions

Backend-specific extension content is not part of the public API.