Interface

DifferentiationInterface.DifferentiationInterface — Module

DifferentiationInterface

An interface to various automatic differentiation backends in Julia.

Exports

AutoChainRules

AutoDiffractor

derivative

gradient

gradient!

handles_input_type

handles_output_type

handles_types

jacobian

jacobian!

multiderivative

multiderivative!

pullback

pullback!

pushforward

pushforward!

value_and_derivative

value_and_gradient

value_and_gradient!

value_and_jacobian

value_and_jacobian!

value_and_multiderivative

value_and_multiderivative!

value_and_pullback

value_and_pullback!

value_and_pushforward

value_and_pushforward!

Utilities

Scalar to scalar

DifferentiationInterface.derivative — Method

derivative(backend, f, x, [extras]) -> der

Compute the derivative der = f'(x) of a scalar-to-scalar function.

DifferentiationInterface.value_and_derivative — Function

value_and_derivative(backend, f, x, [extras]) -> (y, der)

Compute the primal value y = f(x) and the derivative der = f'(x) of a scalar-to-scalar function.

Scalar to array

DifferentiationInterface.multiderivative! — Method

multiderivative!(multider, backend, f, x, [extras]) -> multider

Compute the (array-valued) derivative multider = f'(x) of a scalar-to-array function, overwriting multider if possible.

DifferentiationInterface.multiderivative — Method

multiderivative(backend, f, x, [extras]) -> multider

Compute the (array-valued) derivative multider = f'(x) of a scalar-to-array function.

DifferentiationInterface.value_and_multiderivative! — Function

value_and_multiderivative!(multider, backend, f, x, [extras]) -> (y, multider)

Compute the primal value y = f(x) and the (array-valued) derivative multider = f'(x) of a scalar-to-array function, overwriting multider if possible.

DifferentiationInterface.value_and_multiderivative — Method

value_and_multiderivative(backend, f, x, [extras]) -> (y, multider)

Compute the primal value y = f(x) and the (array-valued) derivative multider = f'(x) of a scalar-to-array function.

Array to scalar

DifferentiationInterface.gradient! — Method

gradient!(grad, backend, f, x, [extras]) -> grad

Compute the gradient grad = ∇f(x) of an array-to-scalar function, overwriting grad if possible.

DifferentiationInterface.gradient — Method

gradient(backend, f, x, [extras]) -> grad

Compute the gradient grad = ∇f(x) of an array-to-scalar function.

DifferentiationInterface.value_and_gradient! — Function

value_and_gradient!(grad, backend, f, x, [extras]) -> (y, grad)

Compute the primal value y = f(x) and the gradient grad = ∇f(x) of an array-to-scalar function, overwriting grad if possible.

DifferentiationInterface.value_and_gradient — Method

value_and_gradient(backend, f, x, [extras]) -> (y, grad)

Compute the primal value y = f(x) and the gradient grad = ∇f(x) of an array-to-scalar function.

Array to array

DifferentiationInterface.jacobian! — Method

jacobian!(jac, backend, f, x, [extras]) -> jac

Compute the Jacobian matrix jac = ∂f(x) of an array-to-array function, overwriting jac if possible.

Notes

Regardless of the shape of x and y, if x has length n and y has length m, then jac is expected to be a m × n matrix. This function acts as if the input and output had been flattened with vec.

DifferentiationInterface.jacobian — Method

jacobian(backend, f, x, [extras]) -> jac

Compute the Jacobian matrix jac = ∂f(x) of an array-to-array function.

Notes

Regardless of the shape of x and y, if x has length n and y has length m, then jac is expected to be a m × n matrix. This function acts as if the input and output had been flattened with vec.

DifferentiationInterface.value_and_jacobian! — Function

value_and_jacobian!(jac, backend, f, x, [extras]) -> (y, jac)

Compute the primal value y = f(x) and the Jacobian matrix jac = ∂f(x) of an array-to-array function, overwriting jac if possible.

Notes

Regardless of the shape of x and y, if x has length n and y has length m, then jac is expected to be a m × n matrix. This function acts as if the input and output had been flattened with vec.

DifferentiationInterface.value_and_jacobian — Method

value_and_jacobian(backend, f, x, [extras]) -> (y, jac)

Compute the primal value y = f(x) and the Jacobian matrix jac = ∂f(x) of an array-to-array function.

Notes

Regardless of the shape of x and y, if x has length n and y has length m, then jac is expected to be a m × n matrix. This function acts as if the input and output had been flattened with vec.

Primitives

Pushforward

DifferentiationInterface.pushforward — Function

pushforward(backend, f, x, dx, [extras]) -> dy

Compute the Jacobian-vector product dy = ∂f(x) * dx.

DifferentiationInterface.pushforward! — Function

pushforward!(dy, backend, f, x, dx, [extras]) -> dy

Compute the Jacobian-vector product dy = ∂f(x) * dx, overwriting dy if possible.

DifferentiationInterface.value_and_pushforward — Function

value_and_pushforward(backend, f, x, dx, [extras]) -> (y, dy)

Compute the primal value y = f(x) and the Jacobian-vector product dy = ∂f(x) * dx.

DifferentiationInterface.value_and_pushforward! — Function

value_and_pushforward!(dy, backend, f, x, dx, [extras]) -> (y, dy)

Compute the primal value y = f(x) and the Jacobian-vector product dy = ∂f(x) * dx, overwriting dy if possible.

Interface requirement

This is the only required implementation for a forward mode backend.

Pullback

DifferentiationInterface.pullback — Function

pullback(backend, f, x, dy, [extras]) -> dx

Compute the vector-Jacobian product dx = ∂f(x)' * dy.

DifferentiationInterface.pullback! — Function

pullback!(dx, backend, f, x, dy, [extras]) -> dx

Compute the vector-Jacobian product dx = ∂f(x)' * dy, overwriting dx if possible.

DifferentiationInterface.value_and_pullback — Function

value_and_pullback(backend, f, x, dy, [extras]) -> (y, dx)

Compute the primal value y = f(x) and the vector-Jacobian product dx = ∂f(x)' * dy.

DifferentiationInterface.value_and_pullback! — Function

value_and_pullback!(dx, backend, f, x, dy, [extras]) -> (y, dx)

Compute the primal value y = f(x) and the vector-Jacobian product dx = ∂f(x)' * dy, overwriting dx if possible.

Interface requirement

This is the only required implementation for a reverse mode backend.

Backends

ADTypes.jl

The following backends are defined by ADTypes.jl:

ADTypes.AbstractADType — Type

Base type for AD choices.

Only a subset is supported by DifferentiationInterface.jl at the moment.

DifferentiationInterface.jl

The following backends are defined by DifferentiationInterface.jl:

DifferentiationInterface.AutoChainRules — Type

AutoChainRules{RC}

Enables the use of AD libraries based on ChainRulesCore.jl.

Fields

ruleconfig::RC: a RuleConfig object

Example

using DifferentiationInterface, Zygote
backend = AutoChainRules(Zygote.ZygoteRuleConfig())

DifferentiationInterface.AutoDiffractor — Type

AutoDiffractor

Enables the use of Diffractor.jl.

Input / output types

DifferentiationInterface.handles_input_type — Method

handles_input_type(backend, ::Type{X})

Check if backend can differentiate functions with input type X.

DifferentiationInterface.handles_output_type — Method

handles_output_type(backend, ::Type{Y})

Check if backend can differentiate functions with output type Y.

DifferentiationInterface.handles_types — Method

handles_types(backend, ::Type{X}, ::Type{Y})

Check if backend can differentiate functions with input type X and output type Y.

Internals

DifferentiationInterface.CustomImplem — Type

CustomImplem

Trait specifying that the custom utilities from the backend should be used as much as possible. Used for internal dispatch only.

DifferentiationInterface.FallbackImplem — Type

FallbackImplem

Trait specifying that the fallback utilities from DifferentiationInterface.jl should be used as much as possible, until they call a pushforward or pullback. Used for internal dispatch only.

DifferentiationInterface.ForwardMode — Type

ForwardMode

Trait identifying forward mode AD backends. Used for internal dispatch only.

DifferentiationInterface.ReverseMode — Type

ReverseMode

Trait identifying reverse mode AD backends. Used for internal dispatch only.

DifferentiationInterface.basisarray — Method

basisarray(backend, a::AbstractArray, i::CartesianIndex)

Construct the i-th stardard basis array in the vector space of a with element type eltype(a).

Note

If an AD backend benefits from a more specialized basis array implementation, this function can be extended on the backend type.

DifferentiationInterface.autodiff_mode — Method

autodiff_mode(backend)

Return ForwardMode() or ReverseMode() in a statically predictable way.

This function must be overloaded for backends that do not inherit from ADTypes.AbstractForwardMode or ADTypes.AbstractReverseMode (e.g. because they support both forward and reverse).

We classify ADTypes.AbstractFiniteDifferencesMode as forward mode.