API reference

DifferentiationInterfaceTest — Module

DifferentiationInterfaceTest

Testing and benchmarking utilities for automatic differentiation in Julia.

Exports

DerivativeScenario

DifferentiationBenchmarkDataRow

GradientScenario

HVPScenario

HessianScenario

JacobianScenario

PullbackScenario

PushforwardScenario

Scenario

SecondDerivativeScenario

benchmark_differentiation

component_scenarios

default_scenarios

gpu_scenarios

sparse_scenarios

static_scenarios

test_differentiation

Entry points

DifferentiationInterfaceTest.test_differentiation — Function

test_differentiation(
    backends::Vector{<:ADTypes.AbstractADType};
    ...
)
test_differentiation(
    backends::Vector{<:ADTypes.AbstractADType},
    scenarios::Vector{<:Scenario};
    correctness,
    type_stability,
    call_count,
    sparsity,
    detailed,
    input_type,
    output_type,
    first_order,
    second_order,
    onearg,
    twoarg,
    inplace,
    outofplace,
    excluded,
    logging,
    isequal,
    isapprox,
    atol,
    rtol
)

Cross-test a list of backends on a list of scenarios, running a variety of different tests.

Default arguments

scenarios::Vector{<:Scenario}: the output of default_scenarios()

Keyword arguments

Testing:

correctness=true: whether to compare the differentiation results with the theoretical values specified in each scenario

type_stability=false: whether to check type stability with JET.jl (thanks to JET.@test_opt)

sparsity: whether to check sparsity of the jacobian / hessian

detailed=false: whether to print a detailed or condensed test log

Filtering:

input_type=Any, output_type=Any: restrict scenario inputs / outputs to subtypes of this

first_order=true, second_order=true: include first order / second order operators

onearg=true, twoarg=true: include one-argument / two-argument functions

inplace=true, outofplace=true: include in-place / out-of-place operators

Options:

logging=false: whether to log progress

isequal=isequal: function used to compare objects exactly, with the standard signature isequal(x, y)

isapprox=isapprox: function used to compare objects approximately, with the standard signature isapprox(x, y; atol, rtol)

atol=0: absolute precision for correctness testing (when comparing to the reference outputs)

rtol=1e-3: relative precision for correctness testing (when comparing to the reference outputs)

test_differentiation(
    backend::ADTypes.AbstractADType,
    args...;
    kwargs...
)

Shortcut for a single backend.

DifferentiationInterfaceTest.benchmark_differentiation — Function

benchmark_differentiation(
    backends::Vector{<:ADTypes.AbstractADType},
    scenarios::Vector{<:Scenario};
    input_type,
    output_type,
    first_order,
    second_order,
    onearg,
    twoarg,
    inplace,
    outofplace,
    excluded,
    logging
) -> DataFrames.DataFrame

Benchmark a list of backends for a list of operators on a list of scenarios.

The object returned is a DataFrames.DataFrame where each column corresponds to a field of DifferentiationBenchmarkDataRow.

The keyword arguments available here have the same meaning as those in test_differentiation.

DifferentiationInterfaceTest.DifferentiationBenchmarkDataRow — Type

DifferentiationBenchmarkDataRow

Ad-hoc storage type for differentiation benchmarking results.

If you have a vector rows::Vector{DifferentiationBenchmarkDataRow}, you can turn it into a DataFrame as follows:

using DataFrames

df = DataFrame(rows)

The resulting DataFrame will have one column for each of the following fields.

Fields

backend::ADTypes.AbstractADType: backend used for benchmarking

scenario::Scenario: scenario used for benchmarking

operator::Symbol: differentiation operator used for benchmarking, e.g. :gradient or :hessian

calls::Int64: number of calls to the differentiated function for one call to the operator

samples::Int64: number of benchmarking samples taken

evals::Int64: number of evaluations used for averaging in each sample

time::Float64: minimum runtime over all samples, in seconds

allocs::Float64: minimum number of allocations over all samples

bytes::Float64: minimum memory allocated over all samples, in bytes

gc_fraction::Float64: minimum fraction of time spent in garbage collection over all samples, between 0.0 and 1.0

compile_fraction::Float64: minimum fraction of time spent compiling over all samples, between 0.0 and 1.0

See the documentation of Chairmarks.jl for more details on the measurement fields.

Pre-made scenario lists

The precise contents of the scenario lists are not part of the API, only their existence.

DifferentiationInterfaceTest.default_scenarios — Function

default_scenarios(rng=Random.default_rng())

Create a vector of Scenarios with standard array types.

DifferentiationInterfaceTest.sparse_scenarios — Function

sparse_scenarios(rng=Random.default_rng())

Create a vector of Scenarios with sparse array types, focused on sparse Jacobians and Hessians.

DifferentiationInterfaceTest.component_scenarios — Function

component_scenarios(rng=Random.default_rng())

Create a vector of Scenarios with component array types from ComponentArrays.jl.

Warning

This function requires ComponentArrays.jl to be loaded (it is implemented in a package extension).

DifferentiationInterfaceTest.gpu_scenarios — Function

gpu_scenarios(rng=Random.default_rng())

Create a vector of Scenarios with GPU array types from JLArrays.jl.

Warning

This function requires JLArrays.jl to be loaded (it is implemented in a package extension).

DifferentiationInterfaceTest.static_scenarios — Function

static_scenarios(rng=Random.default_rng())

Create a vector of Scenarios with static array types from StaticArrays.jl.

Warning

This function requires StaticArrays.jl to be loaded (it is implemented in a package extension).

Scenario types

DifferentiationInterfaceTest.Scenario — Type

Scenario{op,args,pl}

Store a testing scenario composed of a function and its input + output for a given operator.

This generic type should never be used directly: use the specific constructor corresponding to the operator you want to test, or a predefined list of scenarios.

Constructors

PushforwardScenario

PullbackScenario

DerivativeScenario

GradientScenario

JacobianScenario

SecondDerivativeScenario

HVPScenario

HessianScenario

Type parameters

op: one of :pushforward, :pullback, :derivative, :gradient, :jacobian,:second_derivative, :hvp, :hessian

args: either 1 (for f(x) = y) or 2 (for f!(y, x) = nothing)

pl: either :inplace or :outofplace

Fields

f::Any: function f (if args==1) or f! (if args==2) to apply

x::Any: primal input

y::Any: primal output

seed::Any: seed for pushforward, pullback or HVP

res1::Any: first-order result of the operator

res2::Any: second-order result of the operator (when it makes sense)

Note that the res1 and res2 fields are given more meaningful names in the keyword arguments of each specialized constructor. For example:

the keyword grad of GradientScenario becomes res1

the keyword hess of HessianScenario becomes res2, and the keyword grad becomes res1

DifferentiationInterfaceTest.PushforwardScenario — Function

PushforwardScenario(f; x, y, dx, dy, nb_args, place)

Construct a Scenario to test pushforward and its variants.

DifferentiationInterfaceTest.PullbackScenario — Function

PullbackScenario(f; x, y, dy, dx, nb_args, place)

Construct a Scenario to test pullback and its variants.

DifferentiationInterfaceTest.DerivativeScenario — Function

DerivativeScenario(f; x, y, der, nb_args, place)

Construct a Scenario to test derivative and its variants.

DifferentiationInterfaceTest.GradientScenario — Function

GradientScenario(f; x, y, grad, nb_args, place)

Construct a Scenario to test gradient and its variants.

DifferentiationInterfaceTest.JacobianScenario — Function

JacobianScenario(f; x, y, jac, nb_args, place)

Construct a Scenario to test jacobian and its variants.

DifferentiationInterfaceTest.SecondDerivativeScenario — Function

SecondDerivativeScenario(f; x, y, der, der2, nb_args, place)

Construct a Scenario to test second_derivative and its variants.

DifferentiationInterfaceTest.HVPScenario — Function

HVPScenario(f; x, y, dx, grad, dg, nb_args, place)

Construct a Scenario to test hvp and its variants.

DifferentiationInterfaceTest.HessianScenario — Function

HessianScenario(f; x, y, grad, hess, nb_args, place)

Construct a Scenario to test hessian and its variants.

Internals

This is not part of the public API.

DifferentiationInterfaceTest.AutoZeroForward — Type

AutoZeroForward <: ADTypes.AbstractADType

Trivial backend that sets all derivatives to zero. Used in testing and benchmarking.

DifferentiationInterfaceTest.AutoZeroReverse — Type

AutoZeroReverse <: ADTypes.AbstractADType

Trivial backend that sets all derivatives to zero. Used in testing and benchmarking.

DifferentiationInterfaceTest.add_batchified! — Method

add_batchified!(scens::AbstractVector{<:Scenario})

Add batchified versions to scens of its scenarios which support it (pushforward, pullback and HVP).

DifferentiationInterfaceTest.allocfree_scenarios — Function

allocfree_scenarios(rng::AbstractRNG=default_rng())

Create a vector of Scenarios with functions that do not allocate.

Warning

At the moment, some second-order scenarios are excluded.

DifferentiationInterfaceTest.flux_isapprox — Function

flux_isapprox(x, y; atol, rtol)

Approximate comparison function to use in correctness tests with gradients of Flux.jl networks.

DifferentiationInterfaceTest.flux_isequal — Function

flux_isequal(x, y)

Exact comparison function to use in correctness tests with gradients of Flux.jl networks.

DifferentiationInterfaceTest.flux_scenarios — Function

flux_scenarios(rng=Random.default_rng())

Create a vector of Scenarios with neural networks from Flux.jl.

Warning

This function requires Flux.jl and FiniteDifferences.jl to be loaded (it is implemented in a package extension).

Danger

These scenarios are still experimental and not part of the public API. Their ground truth values are computed with finite differences, and thus subject to imprecision.

DifferentiationInterfaceTest.make_closure — Method

make_closure(scen::Scenario)

Return a new Scenario with a modified function f or f! that closes over differentiable data.

DifferentiationInterfaceTest.test_allocfree — Method

test_allocfree(benchmark_data::DataFrame)

Test that every row in benchmark_data which is not a preparation row has zero allocation.