API

DifferentiationInterface — Module

DifferentiationInterface

An interface to various automatic differentiation backends in Julia.

Exports

AutoChainRules

AutoDiffractor

AutoEnzyme

AutoFastDifferentiation

AutoFiniteDiff

AutoFiniteDifferences

AutoForwardDiff

AutoPolyesterForwardDiff

AutoReverseDiff

AutoSparse

AutoSymbolics

AutoTapir

AutoTracker

AutoZygote

DifferentiateWith

GreedyColoringAlgorithm

SecondOrder

check_available

check_hessian

check_twoarg

derivative

derivative!

gradient

gradient!

hessian

hessian!

hvp

hvp!

jacobian

jacobian!

prepare_derivative

prepare_gradient

prepare_hessian

prepare_hvp

prepare_hvp_same_point

prepare_jacobian

prepare_pullback

prepare_pullback_same_point

prepare_pushforward

prepare_pushforward_same_point

prepare_second_derivative

pullback

pullback!

pushforward

pushforward!

second_derivative

second_derivative!

value_and_derivative

value_and_derivative!

value_and_gradient

value_and_gradient!

value_and_jacobian

value_and_jacobian!

value_and_pullback

value_and_pullback!

value_and_pushforward

value_and_pushforward!

First order

Pushforward

DifferentiationInterface.prepare_pushforward — Function

prepare_pushforward(f,     backend, x, dx) -> extras
prepare_pushforward(f!, y, backend, x, dx) -> extras

Create an extras object that can be given to pushforward and its variants.

Warning

If the function changes in any way, the result of preparation will be invalidated, and you will need to run it again. In the two-argument case, y is mutated by f! during preparation.

DifferentiationInterface.prepare_pushforward_same_point — Function

prepare_pushforward_same_point(f,     backend, x, dx) -> extras_same
prepare_pushforward_same_point(f!, y, backend, x, dx) -> extras_same

Create an extras_same object that can be given to pushforward and its variants if they are applied at the same point x.

Warning

If the function or the point changes in any way, the result of preparation will be invalidated, and you will need to run it again. In the two-argument case, y is mutated by f! during preparation.

DifferentiationInterface.pushforward — Function

pushforward(f,     backend, x, dx, [extras]) -> dy
pushforward(f!, y, backend, x, dx, [extras]) -> dy

Compute the pushforward of the function f at point x with seed dx.

DifferentiationInterface.pushforward! — Function

pushforward!(f,     dy, backend, x, dx, [extras]) -> dy
pushforward!(f!, y, dy, backend, x, dx, [extras]) -> dy

Compute the pushforward of the function f at point x with seed dx, overwriting dy.

DifferentiationInterface.value_and_pushforward — Function

value_and_pushforward(f,     backend, x, dx, [extras]) -> (y, dy)
value_and_pushforward(f!, y, backend, x, dx, [extras]) -> (y, dy)

Compute the value and the pushforward of the function f at point x with seed dx.

Info

Required primitive for forward mode backends.

DifferentiationInterface.value_and_pushforward! — Function

value_and_pushforward!(f,     dy, backend, x, dx, [extras]) -> (y, dy)
value_and_pushforward!(f!, y, dy, backend, x, dx, [extras]) -> (y, dy)

Compute the value and the pushforward of the function f at point x with seed dx, overwriting dy.

Pullback

DifferentiationInterface.prepare_pullback — Function

prepare_pullback(f,     backend, x, dy) -> extras
prepare_pullback(f!, y, backend, x, dy) -> extras

Create an extras object that can be given to pullback and its variants.

Warning

If the function changes in any way, the result of preparation will be invalidated, and you will need to run it again. In the two-argument case, y is mutated by f! during preparation.

DifferentiationInterface.prepare_pullback_same_point — Function

prepare_pullback_same_point(f,     backend, x, dy) -> extras_same
prepare_pullback_same_point(f!, y, backend, x, dy) -> extras_same

Create an extras_same object that can be given to pullback and its variants if they are applied at the same point x.

Warning

If the function or the point changes in any way, the result of preparation will be invalidated, and you will need to run it again. In the two-argument case, y is mutated by f! during preparation.

DifferentiationInterface.pullback — Function

pullback(f,     backend, x, dy, [extras]) -> dx
pullback(f!, y, backend, x, dy, [extras]) -> dx

Compute the pullback of the function f at point x with seed dy.

DifferentiationInterface.pullback! — Function

pullback!(f,     dx, backend, x, dy, [extras]) -> dx
pullback!(f!, y, dx, backend, x, dy, [extras]) -> dx

Compute the pullback of the function f at point x with seed dy, overwriting dx.

DifferentiationInterface.value_and_pullback — Function

value_and_pullback(f,     backend, x, dy, [extras]) -> (y, dx)
value_and_pullback(f!, y, backend, x, dy, [extras]) -> (y, dx)

Compute the value and the pullback of the function f at point x with seed dy.

Info

Required primitive for reverse mode backends.

DifferentiationInterface.value_and_pullback! — Function

value_and_pullback!(f,     dx, backend, x, dy, [extras]) -> (y, dx)
value_and_pullback!(f!, y, dx, backend, x, dy, [extras]) -> (y, dx)

Compute the value and the pullback of the function f at point x with seed dy, overwriting dx.

Derivative

DifferentiationInterface.prepare_derivative — Function

prepare_derivative(f,     backend, x) -> extras
prepare_derivative(f!, y, backend, x) -> extras

Create an extras object that can be given to derivative and its variants.

Warning

If the function changes in any way, the result of preparation will be invalidated, and you will need to run it again. In the two-argument case, y is mutated by f! during preparation.

DifferentiationInterface.derivative — Function

derivative(f,     backend, x, [extras]) -> der
derivative(f!, y, backend, x, [extras]) -> der

Compute the derivative of the function f at point x.

DifferentiationInterface.derivative! — Function

derivative!(f,     der, backend, x, [extras]) -> der
derivative!(f!, y, der, backend, x, [extras]) -> der

Compute the derivative of the function f at point x, overwriting der.

DifferentiationInterface.value_and_derivative — Function

value_and_derivative(f,     backend, x, [extras]) -> (y, der)
value_and_derivative(f!, y, backend, x, [extras]) -> (y, der)

Compute the value and the derivative of the function f at point x.

DifferentiationInterface.value_and_derivative! — Function

value_and_derivative!(f,     der, backend, x, [extras]) -> (y, der)
value_and_derivative!(f!, y, der, backend, x, [extras]) -> (y, der)

Compute the value and the derivative of the function f at point x, overwriting der.

Gradient

DifferentiationInterface.prepare_gradient — Function

prepare_gradient(f, backend, x) -> extras

Create an extras object that can be given to gradient and its variants.

Warning

If the function changes in any way, the result of preparation will be invalidated, and you will need to run it again.

DifferentiationInterface.gradient — Function

gradient(f, backend, x, [extras]) -> grad

Compute the gradient of the function f at point x.

DifferentiationInterface.gradient! — Function

gradient!(f, grad, backend, x, [extras]) -> grad

Compute the gradient of the function f at point x, overwriting grad.

DifferentiationInterface.value_and_gradient — Function

value_and_gradient(f, backend, x, [extras]) -> (y, grad)

Compute the value and the gradient of the function f at point x.

DifferentiationInterface.value_and_gradient! — Function

value_and_gradient!(f, grad, backend, x, [extras]) -> (y, grad)

Compute the value and the gradient of the function f at point x, overwriting grad.

Jacobian

DifferentiationInterface.prepare_jacobian — Function

prepare_jacobian(f,     backend, x) -> extras
prepare_jacobian(f!, y, backend, x) -> extras

Create an extras object that can be given to jacobian and its variants.

Warning

If the function changes in any way, the result of preparation will be invalidated, and you will need to run it again. In the two-argument case, y is mutated by f! during preparation.

DifferentiationInterface.jacobian — Function

jacobian(f,     backend, x, [extras]) -> jac
jacobian(f!, y, backend, x, [extras]) -> jac

Compute the Jacobian matrix of the function f at point x.

DifferentiationInterface.jacobian! — Function

jacobian!(f,     jac, backend, x, [extras]) -> jac
jacobian!(f!, y, jac, backend, x, [extras]) -> jac

Compute the Jacobian matrix of the function f at point x, overwriting jac.

DifferentiationInterface.value_and_jacobian — Function

value_and_jacobian(f,     backend, x, [extras]) -> (y, jac)
value_and_jacobian(f!, y, backend, x, [extras]) -> (y, jac)

Compute the value and the Jacobian matrix of the function f at point x.

DifferentiationInterface.value_and_jacobian! — Function

value_and_jacobian!(f,     jac, backend, x, [extras]) -> (y, jac)
value_and_jacobian!(f!, y, jac, backend, x, [extras]) -> (y, jac)

Compute the value and the Jacobian matrix of the function f at point x, overwriting jac.

Second order

DifferentiationInterface.SecondOrder — Type

SecondOrder

Combination of two backends for second-order differentiation.

Danger

SecondOrder backends do not support first-order operators.

Constructor

SecondOrder(outer_backend, inner_backend)

Fields

outer::ADTypes.AbstractADType: backend for the outer differentiation

inner::ADTypes.AbstractADType: backend for the inner differentiation

Second derivative

DifferentiationInterface.prepare_second_derivative — Function

prepare_second_derivative(f, backend, x) -> extras

Create an extras object that can be given to second_derivative and its variants.

Warning

If the function changes in any way, the result of preparation will be invalidated, and you will need to run it again.

DifferentiationInterface.second_derivative — Function

second_derivative(f, backend, x, [extras]) -> der2

Compute the second derivative of the function f at point x.

DifferentiationInterface.second_derivative! — Function

second_derivative!(f, der2, backend, x, [extras]) -> der2

Compute the second derivative of the function f at point x, overwriting der2.

Hessian-vector product

DifferentiationInterface.prepare_hvp — Function

prepare_hvp(f, backend, x, v) -> extras

Create an extras object that can be given to hvp and its variants.

Warning

If the function changes in any way, the result of preparation will be invalidated, and you will need to run it again.

DifferentiationInterface.prepare_hvp_same_point — Function

prepare_hvp_same_point(f, backend, x, v) -> extras_same

Create an extras_same object that can be given to hvp and its variants if they are applied at the same point x.

Warning

If the function or the point changes in any way, the result of preparation will be invalidated, and you will need to run it again.

DifferentiationInterface.hvp — Function

hvp(f, backend, x, v, [extras]) -> p

Compute the Hessian-vector product of f at point x with seed v.

DifferentiationInterface.hvp! — Function

hvp!(f, p, backend, x, v, [extras]) -> p

Compute the Hessian-vector product of f at point x with seed v, overwriting p.

Hessian

DifferentiationInterface.prepare_hessian — Function

prepare_hessian(f, backend, x) -> extras

Create an extras object that can be given to hessian and its variants.

Warning

If the function changes in any way, the result of preparation will be invalidated, and you will need to run it again.

DifferentiationInterface.hessian — Function

hessian(f, backend, x, [extras]) -> hess

Compute the Hessian matrix of the function f at point x.

DifferentiationInterface.hessian! — Function

hessian!(f, hess, backend, x, [extras]) -> hess

Compute the Hessian matrix of the function f at point x, overwriting hess.

Utilities

Backend queries

DifferentiationInterface.check_available — Function

check_available(backend)

Check whether backend is available (i.e. whether the extension is loaded).

DifferentiationInterface.check_twoarg — Function

check_twoarg(backend)

Check whether backend supports differentiation of two-argument functions.

DifferentiationInterface.check_hessian — Function

check_hessian(backend)

Check whether backend supports second order differentiation by trying to compute a hessian.

Warning

Might take a while due to compilation time.

DifferentiationInterface.outer — Function

outer(backend::SecondOrder)

Return the outer backend of a SecondOrder object, tasked with differentiation at the second order.

DifferentiationInterface.inner — Function

inner(backend::SecondOrder)

Return the inner backend of a SecondOrder object, tasked with differentiation at the first order.

Backend switch

DifferentiationInterface.DifferentiateWith — Type

DifferentiateWith

Callable function wrapper that enforces differentiation with a specified (inner) backend.

This works by defining new rules overriding the behavior of the outer backend that would normally be used.

Warning

This is an experimental functionality, whose API cannot yet be considered stable. At the moment, it only supports one-argument functions, and rules are only defined for ChainRules.jl-compatible outer backends.

Fields

f: the function in question

backend::AbstractADType: the inner backend to use for differentiation

Constructor

DifferentiateWith(f, backend)

Example

using DifferentiationInterface
import ForwardDiff, Zygote

function f(x)
    a = Vector{eltype(x)}(undef, 1)
    a[1] = sum(x)  # mutation that breaks Zygote
    return a[1]
end

dw = DifferentiateWith(f, AutoForwardDiff());

gradient(dw, AutoZygote(), [2.0])  # calls ForwardDiff instead

# output

1-element Vector{Float64}:
 1.0

Internals

The following is not part of the public API.

DifferentiationInterface.DerivativeExtras — Type

DerivativeExtras

Abstract type for additional information needed by derivative and its variants.

DifferentiationInterface.ForwardOverForward — Type

ForwardOverForward

Traits identifying second-order backends that compute HVPs in forward over forward mode (inefficient).

DifferentiationInterface.ForwardOverReverse — Type

ForwardOverReverse

Traits identifying second-order backends that compute HVPs in forward over reverse mode.

DifferentiationInterface.GradientExtras — Type

GradientExtras

Abstract type for additional information needed by gradient and its variants.

DifferentiationInterface.HVPExtras — Type

HVPExtras

Abstract type for additional information needed by hvp and its variants.

DifferentiationInterface.HessianExtras — Type

HessianExtras

Abstract type for additional information needed by hessian and its variants.

DifferentiationInterface.JacobianExtras — Type

JacobianExtras

Abstract type for additional information needed by jacobian and its variants.

DifferentiationInterface.PullbackExtras — Type

PullbackExtras

Abstract type for additional information needed by pullback and its variants.

DifferentiationInterface.PullbackFast — Type

PullbackFast

Trait identifying backends that support efficient pullbacks.

DifferentiationInterface.PullbackSlow — Type

PullbackSlow

Trait identifying backends that do not support efficient pullbacks.

DifferentiationInterface.PushforwardExtras — Type

PushforwardExtras

Abstract type for additional information needed by pushforward and its variants.

DifferentiationInterface.PushforwardFast — Type

PushforwardFast

Trait identifying backends that support efficient pushforwards.

DifferentiationInterface.PushforwardSlow — Type

PushforwardSlow

Trait identifying backends that do not support efficient pushforwards.

DifferentiationInterface.ReverseOverForward — Type

ReverseOverForward

Traits identifying second-order backends that compute HVPs in reverse over forward mode.

DifferentiationInterface.ReverseOverReverse — Type

ReverseOverReverse

Traits identifying second-order backends that compute HVPs in reverse over reverse mode.

DifferentiationInterface.SecondDerivativeExtras — Type

SecondDerivativeExtras

Abstract type for additional information needed by second_derivative and its variants.

DifferentiationInterface.SymbolicsSparsityDetector — Type

SymbolicsSparsityDetector <: ADTypes.AbstractSparsityDetector

Sparsity detection algorithm based on the Symbolics.jl tracing system.

Compatible with the ADTypes.jl sparsity detection framework.

Danger

This functionality is in a package extension, and requires Symbolics.jl to be loaded.

Implements

ADTypes.jacobian_sparsity

ADTypes.hessian_sparsity

Reference

Sparsity Programming: Automated Sparsity-Aware Optimizations in Differentiable Programming, Gowda et al. (2019)

DifferentiationInterface.TwoArgNotSupported — Type

TwoArgNotSupported

Trait identifying backends that do not support two-argument functions f!(y, x).

DifferentiationInterface.TwoArgSupported — Type

TwoArgSupported

Trait identifying backends that support two-argument functions f!(y, x).

ADTypes.mode — Method

mode(backend::SecondOrder)

Return the outer mode of the second-order backend.

DifferentiationInterface.basis — Method

basis(backend, a::AbstractArray, i::CartesianIndex)

Construct the i-th stardard basis array in the vector space of a with element type eltype(a).

Note

If an AD backend benefits from a more specialized basis array implementation, this function can be extended on the backend type.

DifferentiationInterface.col_major — Method

col_major(A::AbstractMatrix)

Construct a column-major representation of the matrix A.

DifferentiationInterface.pick_chunksize — Method

pick_chunksize(input_length)

Pick a reasonable chunk size for chunked derivative evaluation with an input of length input_length.

The result cannot be larger than DEFAULT_CHUNKSIZE=8.

DifferentiationInterface.pullback_performance — Method

pullback_performance(backend)

Return PullbackFast or PullbackSlow in a statically predictable way.

DifferentiationInterface.pushforward_performance — Method

pushforward_performance(backend)

Return PushforwardFast or PushforwardSlow in a statically predictable way.

DifferentiationInterface.row_major — Method

row_major(A::AbstractMatrix)

Construct a row-major representation of the matrix A.

DifferentiationInterface.twoarg_support — Method

twoarg_support(backend)

Return TwoArgSupported or TwoArgNotSupported in a statically predictable way.