API reference

DifferentiationInterfaceTest — Module

DifferentiationInterfaceTest

Testing and benchmarking utilities for automatic differentiation in Julia.

Exports

AbstractScenario

DerivativeScenario

GradientScenario

HVPScenario

HessianScenario

JacobianScenario

PullbackScenario

PushforwardScenario

SecondDerivativeScenario

benchmark_differentiation

component_scenarios

default_scenarios

gpu_scenarios

static_scenarios

test_differentiation

Entry points

DifferentiationInterfaceTest.test_differentiation — Function

test_differentiation(
    backends::Vector{<:ADTypes.AbstractADType};
    ...
)
test_differentiation(
    backends::Vector{<:ADTypes.AbstractADType},
    scenarios::Vector{<:AbstractScenario};
    correctness,
    type_stability,
    call_count,
    sparsity,
    ref_backend,
    detailed,
    input_type,
    output_type,
    first_order,
    second_order,
    excluded,
    logging,
    isapprox,
    atol,
    rtol
)

Cross-test a list of backends on a list of scenarios, running a variety of different tests.

Default arguments

scenarios::Vector{<:AbstractScenario}: the output of default_scenarios()

Keyword arguments

Testing:

correctness=true: whether to compare the differentiation results with the theoretical values specified in each scenario

type_stability=false: whether to check type stability with JET.jl (thanks to @test_opt)

sparsity: whether to check sparsity of the jacobian / hessian

ref_backend: if not nothing, an ADTypes.AbstractADType object to use instead of the scenario-specific reference to provide true values

detailed=false: whether to print a detailed or condensed test log

Filtering:

input_type=Any: restrict scenario inputs to subtypes of this

output_type=Any: restrict scenario outputs to subtypes of this

first_order=true: include first order operators

second_order=true: include second order operators

Options:

logging=false: whether to log progress

isapprox=isapprox: function used to compare objects, with the standard signature isapprox(x, y; atol, rtol)

atol=0: absolute precision for correctness testing (when comparing to the reference outputs)

rtol=1e-3: relative precision for correctness testing (when comparing to the reference outputs)

test_differentiation(
    backend::ADTypes.AbstractADType,
    args...;
    kwargs...
)

Shortcut for a single backend.

DifferentiationInterfaceTest.benchmark_differentiation — Function

benchmark_differentiation(
    backends::Vector{<:ADTypes.AbstractADType};
    ...
) -> Vector{DifferentiationInterfaceTest.BenchmarkDataRow}
benchmark_differentiation(
    backends::Vector{<:ADTypes.AbstractADType},
    scenarios::Vector{<:AbstractScenario};
    input_type,
    output_type,
    first_order,
    second_order,
    excluded,
    logging
) -> Vector{DifferentiationInterfaceTest.BenchmarkDataRow}

Benchmark a list of backends for a list of operators on a list of scenarios.

Keyword arguments

filtering: same as test_differentiation for the filtering part.

logging=false: whether to log progress

Pre-made scenario lists

DifferentiationInterfaceTest.default_scenarios — Function

default_scenarios()

Create a vector of AbstractScenarios with standard array types.

DifferentiationInterfaceTest.component_scenarios — Function

component_scenarios()

Create a vector of AbstractScenarios with component array types from ComponentArrays.jl.

DifferentiationInterfaceTest.gpu_scenarios — Function

gpu_scenarios()

Create a vector of AbstractScenarios with GPU array types from JLArrays.jl.

DifferentiationInterfaceTest.static_scenarios — Function

static_scenarios()

Create a vector of AbstractScenarios with static array types from StaticArrays.jl.

Scenario types

DifferentiationInterfaceTest.AbstractScenario — Type

AbstractScenario

Store a testing scenario composed of a function and its input + output.

This abstract type should never be used directly: construct one of the subtypes corresponding to the operator you want to test.

Subtypes

PushforwardScenario

PullbackScenario

DerivativeScenario

GradientScenario

JacobianScenario

SecondDerivativeScenario

HVPScenario

HessianScenario

Fields

All subtypes have the following fields:

f: function to apply

x: primal input

y: primal output

ref: reference to compare against

In addition, some subtypes contain an additional seed (dx or dy).

Constructor

If y is provided, f is interpreted as a 2-argument function f!(y, x) = nothing. Otherwise, f is interpreted as an 1-argument function f(x) = y.

The reference keyword ref should be a function that takes x (and a potential seed dx or dy) to return the correct object.

The operator behavior keyword operator should be either :inplace or :outofplace depending on what must be tested.

DifferentiationInterfaceTest.PushforwardScenario — Type

PushforwardScenario(f; x, y, dx, ref, operator)

See AbstractScenario for details.

DifferentiationInterfaceTest.PullbackScenario — Type

PullbackScenario(f; x, y, dy, ref, operator)

See AbstractScenario for details.

DifferentiationInterfaceTest.DerivativeScenario — Type

DerivativeScenario(f; x, y, ref, operator)

See AbstractScenario for details.

DifferentiationInterfaceTest.GradientScenario — Type

GradientScenario(f; x, y, ref, operator)

See AbstractScenario for details.

DifferentiationInterfaceTest.JacobianScenario — Type

JacobianScenario(f; x, y, ref, operator)

See AbstractScenario for details.

DifferentiationInterfaceTest.SecondDerivativeScenario — Type

SecondDerivativeScenario(f; x, y, ref, operator)

See AbstractScenario for details.

DifferentiationInterfaceTest.HVPScenario — Type

HVPScenario(f; x, y, dx, ref, operator)

See AbstractScenario for details.

DifferentiationInterfaceTest.HessianScenario — Type

HessianScenario(f; x, y, ref, operator)

See AbstractScenario for details.

Internals

This is not part of the public API.

DifferentiationInterfaceTest.AutoZeroForward — Type

AutoZeroForward <: ADTypes.AbstractForwardMode

Trivial backend that sets all derivatives to zero. Used in testing and benchmarking.

DifferentiationInterfaceTest.AutoZeroReverse — Type

AutoZeroReverse <: ADTypes.AbstractReverseMode

Trivial backend that sets all derivatives to zero. Used in testing and benchmarking.

DifferentiationInterfaceTest.BenchmarkDataRow — Type

BenchmarkDataRow

Ad-hoc storage type for differentiation benchmarking results. If you have a vector rows::Vector{BenchmarkDataRow}, you can turn it into a DataFrame as follows:

df = DataFrames.DataFrame(rows)

Fields

These are not part of the public API.

backend::String

mode::Type

scenario::Symbol

operator::Symbol

func::Symbol

arguments::Int64

input_type::Type

output_type::Type

input_size::Tuple

output_size::Tuple

calls::Int64

samples::Int64

evals::Int64

time::Float64

bytes::Float64

allocs::Float64

compile_fraction::Float64

gc_fraction::Float64