API reference

DifferentiationInterface — Module

DifferentiationInterface

An interface to various automatic differentiation backends in Julia.

Exports

SecondOrder

check_available

check_hessian

check_twoarg

derivative

derivative!

gradient

gradient!

hessian

hessian!

hvp

hvp!

jacobian

jacobian!

prepare_derivative

prepare_gradient

prepare_hessian

prepare_hvp

prepare_jacobian

prepare_pullback

prepare_pushforward

prepare_second_derivative

pullback

pullback!

pushforward

pushforward!

second_derivative

second_derivative!

value_and_derivative

value_and_derivative!

value_and_gradient

value_and_gradient!

value_and_jacobian

value_and_jacobian!

value_and_pullback

value_and_pullback!

value_and_pullback!_split

value_and_pullback_split

value_and_pushforward

value_and_pushforward!

Derivative

DifferentiationInterface.prepare_derivative — Function

prepare_derivative(f,     backend, x) -> extras
prepare_derivative(f!, y, backend, x) -> extras

Create an extras object subtyping DerivativeExtras that can be given to derivative operators.

Beware that in the two-argument case, y is mutated by f! during preparation.

DifferentiationInterface.derivative — Function

derivative(f,     backend, x, [extras]) -> der
derivative(f!, y, backend, x, [extras]) -> der

DifferentiationInterface.derivative! — Function

derivative!(f,     der, backend, x, [extras]) -> der
derivative!(f!, y, der, backend, x, [extras]) -> der

DifferentiationInterface.value_and_derivative — Function

value_and_derivative(f,     backend, x, [extras]) -> (y, der)
value_and_derivative(f!, y, backend, x, [extras]) -> (y, der)

DifferentiationInterface.value_and_derivative! — Function

value_and_derivative!(f,     der, backend, x, [extras]) -> (y, der)
value_and_derivative!(f!, y, der, backend, x, [extras]) -> (y, der)

Gradient

DifferentiationInterface.prepare_gradient — Function

prepare_gradient(f, backend, x) -> extras

Create an extras object subtyping GradientExtras that can be given to gradient operators.

DifferentiationInterface.gradient — Function

gradient(f, backend, x, [extras]) -> grad

DifferentiationInterface.gradient! — Function

gradient!(f, grad, backend, x, [extras]) -> grad

DifferentiationInterface.value_and_gradient — Function

value_and_gradient(f, backend, x, [extras]) -> (y, grad)

DifferentiationInterface.value_and_gradient! — Function

value_and_gradient!(f, grad, backend, x, [extras]) -> (y, grad)

Jacobian

DifferentiationInterface.prepare_jacobian — Function

prepare_jacobian(f,     backend, x) -> extras
prepare_jacobian(f!, y, backend, x) -> extras

Create an extras object subtyping JacobianExtras that can be given to Jacobian operators.

Beware that in the two-argument case, y is mutated by f! during preparation.

DifferentiationInterface.jacobian — Function

jacobian(f,     backend, x, [extras]) -> jac
jacobian(f!, y, backend, x, [extras]) -> jac

DifferentiationInterface.jacobian! — Function

jacobian!(f,     jac, backend, x, [extras]) -> jac
jacobian!(f!, y, jac, backend, x, [extras]) -> jac

DifferentiationInterface.value_and_jacobian — Function

value_and_jacobian(f,     backend, x, [extras]) -> (y, jac)
value_and_jacobian(f!, y, backend, x, [extras]) -> (y, jac)

DifferentiationInterface.value_and_jacobian! — Function

value_and_jacobian!(f,     jac, backend, x, [extras]) -> (y, jac)
value_and_jacobian!(f!, y, jac, backend, x, [extras]) -> (y, jac)

Second order

DifferentiationInterface.SecondOrder — Type

SecondOrder

Combination of two backends for second-order differentiation.

Constructor

SecondOrder(outer, inner)

Fields

outer::AbstractADType: backend for the outer differentiation

inner::AbstractADType: backend for the inner differentiation

DifferentiationInterface.prepare_second_derivative — Function

prepare_second_derivative(f, backend, x) -> extras

Create an extras object subtyping SecondDerivativeExtras that can be given to second derivative operators.

DifferentiationInterface.second_derivative — Function

second_derivative(f, backend, x, [extras]) -> der2

DifferentiationInterface.second_derivative! — Function

second_derivative!(f, der2, backend, x, [extras]) -> der2

DifferentiationInterface.prepare_hvp — Function

prepare_hvp(f, backend, x, v) -> extras

Create an extras object subtyping HVPExtras that can be given to Hessian-vector product operators.

DifferentiationInterface.hvp — Function

hvp(f, backend, x, v, [extras]) -> p

DifferentiationInterface.hvp! — Function

hvp!(f, p, backend, x, v, [extras]) -> p

DifferentiationInterface.prepare_hessian — Function

prepare_hessian(f, backend, x) -> extras

Create an extras object subtyping HessianExtras that can be given to Hessian operators.

DifferentiationInterface.hessian — Function

hessian(f, backend, x, [extras]) -> hess

DifferentiationInterface.hessian! — Function

hessian!(f, hess, backend, x, [extras]) -> hess

Primitives

DifferentiationInterface.prepare_pushforward — Function

prepare_pushforward(f,     backend, x, dx) -> extras
prepare_pushforward(f!, y, backend, x, dx) -> extras

Create an extras object subtyping PushforwardExtras that can be given to pushforward operators.

Beware that in the two-argument case, y is mutated by f! during preparation.

DifferentiationInterface.pushforward — Function

pushforward(f,     backend, x, dx, [extras]) -> dy
pushforward(f!, y, backend, x, dx, [extras]) -> dy

DifferentiationInterface.pushforward! — Function

pushforward!(f,     dy, backend, x, dx, [extras]) -> dy
pushforward!(f!, y, dy, backend, x, dx, [extras]) -> dy

DifferentiationInterface.value_and_pushforward — Function

value_and_pushforward(f,     backend, x, dx, [extras]) -> (y, dy)
value_and_pushforward(f!, y, backend, x, dx, [extras]) -> (y, dy)

Info

Required primitive for forward mode backends.

DifferentiationInterface.value_and_pushforward! — Function

value_and_pushforward!(f,     dy, backend, x, dx, [extras]) -> (y, dy)
value_and_pushforward!(f!, y, dy, backend, x, dx, [extras]) -> (y, dy)

DifferentiationInterface.prepare_pullback — Function

prepare_pullback(f,     backend, x, dy) -> extras
prepare_pullback(f!, y, backend, x, dy) -> extras

Create an extras object subtyping PullbackExtras that can be given to pullback operators.

Beware that in the two-argument case, y is mutated by f! during preparation.

DifferentiationInterface.pullback — Function

pullback(f,     backend, x, dy, [extras]) -> dx
pullback(f!, y, backend, x, dy, [extras]) -> dx

DifferentiationInterface.pullback! — Function

pullback!(f,     dx, backend, x, dy, [extras]) -> dx
pullback!(f!, y, dx, backend, x, dy, [extras]) -> dx

DifferentiationInterface.value_and_pullback — Function

value_and_pullback(f,     backend, x, dy, [extras]) -> (y, dx)
value_and_pullback(f!, y, backend, x, dy, [extras]) -> (y, dx)

Info

Required primitive for reverse mode backends.

DifferentiationInterface.value_and_pullback! — Function

value_and_pullback!(f,     dx, backend, x, dy, [extras]) -> (y, dx)
value_and_pullback!(f!, y, dx, backend, x, dy, [extras]) -> (y, dx)

DifferentiationInterface.value_and_pullback_split — Function

value_and_pullback_split(f,     backend, x, [extras]) -> (y, pbf(   dy) -> dx)
value_and_pullback_split(f!, y, backend, x, [extras]) -> (y, pbf(y, dy) -> dx)

DifferentiationInterface.value_and_pullback!_split — Function

value_and_pullback!_split(f,     backend, x, [extras]) -> (y, pbf!(   dx, dy) -> dx)
value_and_pullback!_split(f!, y, backend, x, [extras]) -> (y, pbf!(y, dx, dy) -> dx)

Backend queries

DifferentiationInterface.check_available — Function

check_available(backend)

Check whether backend is available (i.e. whether the extension is loaded).

DifferentiationInterface.check_twoarg — Function

check_twoarg(backend)

Check whether backend supports differentiation of two-argument functions.

DifferentiationInterface.check_hessian — Function

check_hessian(backend)

Check whether backend supports second order differentiation by trying to compute a hessian.

Warning

Might take a while due to compilation time.

Internals

This is not part of the public API.

DifferentiationInterface.AdjacencyGraph — Type

AdjacencyGraph

Represent a graph between the columns of a symmetric n × n matrix A with nonzero diagonal elements.

This graph is defined as G = (C, E) where C = 1:n is the set of columns and (i, j) ∈ E whenever A[i, j] is nonzero for some j ∈ 1:m, j ≠ i.

Fields

A::AbstractMatrix

Reference

What Color Is Your Jacobian? Graph Coloring for Computing Derivatives, Gebremedhin et al. (2005)

DifferentiationInterface.BipartiteGraph — Type

BipartiteGraph

Represent a bipartite graph between the rows and the columns of a non-symmetric m × n matrix A.

This graph is defined as G = (R, C, E) where R = 1:m is the set of row indices, C = 1:n is the set of column indices and (i, j) ∈ E whenever A[i, j] is nonzero.

Fields

A::AbstractMatrix

Reference

What Color Is Your Jacobian? Graph Coloring for Computing Derivatives, Gebremedhin et al. (2005)

DifferentiationInterface.CompressedMatrix — Type

CompressedMatrix{dir}

Compressed representation B of a sparse matrix A ∈ ℝ^{m×n} obtained by summing some of its columns (if dir == :col) or rows (if dir == :row), grouped by color.

Fields

sparsity::AbstractMatrix{Bool}: boolean sparsity pattern of the matrix A

colors::Vector{Int}: vector such that

if dir ==:col, thencolors[j] ∈ 1:cis the color of columnj`

if dir ==:row, thencolors[i] ∈ 1:cis the color of rowi`

groups::Vector{Vector{Int}}: vector of length c such that

if dir == :col, then groups[k] is the vector of column indices assigned to the same color k ∈ 1:c

if dir == :row, then groups[k] is the vector of row indices assigned to the same color k ∈ 1:c

aggregates::AbstractMatrix: matrix B such that

if dir == :col, then size(B) = (m, c) and B[:, c] = sum(A[:, k] for k in groups[c])

if dir == :row, then size(B) = (c, n) and B[c, :] = sum(A[k, :] for k in groups[c])

DifferentiationInterface.CompressedMatrix — Method

CompressedMatrix{dir}(sparsity, groups, aggregates)

Constructor for CompressedMatrix.

DifferentiationInterface.DerivativeExtras — Type

DerivativeExtras

Abstract type for additional information needed by derivative operators.

DifferentiationInterface.ForwardOverForward — Type

ForwardOverForward

Traits identifying second-order backends that compute HVPs in forward over forward mode (inefficient).

DifferentiationInterface.ForwardOverReverse — Type

ForwardOverReverse

Traits identifying second-order backends that compute HVPs in forward over reverse mode.

DifferentiationInterface.GradientExtras — Type

GradientExtras

Abstract type for additional information needed by gradient operators.

DifferentiationInterface.GreedyColoringAlgorithm — Type

GreedyColoringAlgorithm <: ADTypes.AbstractColoringAlgorithm

Matrix coloring algorithm for sparse Jacobians and Hessians.

Compatible with the ADTypes.jl coloring framework.

See also

ADTypes.column_coloring

ADTypes.row_coloring

ADTypes.symmetric_coloring

Reference

What Color Is Your Jacobian? Graph Coloring for Computing Derivatives, Gebremedhin et al. (2005)

DifferentiationInterface.HVPExtras — Type

HVPExtras

Abstract type for additional information needed by Hessian-vector product operators.

DifferentiationInterface.HessianExtras — Type

HessianExtras

Abstract type for additional information needed by Hessian operators.

DifferentiationInterface.JacobianExtras — Type

JacobianExtras

Abstract type for additional information needed by Jacobian operators.

DifferentiationInterface.MutationNotSupported — Type

MutationNotSupported

Trait identifying backends that do not support two-argument functions f!(y, x).

DifferentiationInterface.MutationSupported — Type

MutationSupported

Trait identifying backends that support two-argument functions f!(y, x).

DifferentiationInterface.PullbackExtras — Type

PullbackExtras

Abstract type for additional information needed by pullback operators.

DifferentiationInterface.PullbackFast — Type

PullbackFast

Trait identifying backends that support efficient pullbacks.

DifferentiationInterface.PullbackSlow — Type

PullbackSlow

Trait identifying backends that do not support efficient pullbacks.

DifferentiationInterface.PushforwardExtras — Type

PushforwardExtras

Abstract type for additional information needed by pushforward operators.

DifferentiationInterface.PushforwardFast — Type

PushforwardFast

Trait identifying backends that support efficient pushforwards.

DifferentiationInterface.PushforwardSlow — Type

PushforwardSlow

Trait identifying backends that do not support efficient pushforwards.

DifferentiationInterface.ReverseOverForward — Type

ReverseOverForward

Traits identifying second-order backends that compute HVPs in reverse over forward mode.

DifferentiationInterface.ReverseOverReverse — Type

ReverseOverReverse

Traits identifying second-order backends that compute HVPs in reverse over reverse mode.

DifferentiationInterface.SecondDerivativeExtras — Type

SecondDerivativeExtras

Abstract type for additional information needed by second derivative operators.

DifferentiationInterface.SymbolicsSparsityDetector — Type

SymbolicsSparsityDetector <: ADTypes.AbstractSparsityDetector

Sparsity detection algorithm based on the Symbolics.jl tracing system.

Compatible with the ADTypes.jl sparsity detection framework.

Danger

This functionality is implemented in an extension, and requires Symbolics.jl to be loaded.

See also

ADTypes.jacobian_sparsity

ADTypes.hessian_sparsity

Reference

Sparsity Programming: Automated Sparsity-Aware Optimizations in Differentiable Programming, Gowda et al. (2019)

ADTypes.mode — Method

mode(backend::SecondOrder)

Return the outer mode of the second-order backend.

DifferentiationInterface.basis — Method

basis(backend, a::AbstractArray, i::CartesianIndex)

Construct the i-th stardard basis array in the vector space of a with element type eltype(a).

Note

If an AD backend benefits from a more specialized basis array implementation, this function can be extended on the backend type.

DifferentiationInterface.mutation_support — Method

mutation_support(backend)

Return MutationSupported or MutationNotSupported in a statically predictable way.

DifferentiationInterface.pick_chunksize — Method

pick_chunksize(input_length)

Pick a reasonable chunk size for chunked derivative evaluation with an input of length input_length.

The result cannot be larger than DEFAULT_CHUNKSIZE=8.

DifferentiationInterface.pullback_performance — Method

pullback_performance(backend)

Return PullbackFast or PullbackSlow in a statically predictable way.

DifferentiationInterface.pushforward_performance — Method

pushforward_performance(backend)

Return PushforwardFast or PushforwardSlow in a statically predictable way.