API reference

DifferentiationInterface.DifferentiationInterface — Module

DifferentiationInterface

An experimental redesign for AbstractDifferentiation.jl.

Exports

ChainRulesForwardBackend

ChainRulesReverseBackend

EnzymeForwardBackend

EnzymeReverseBackend

FiniteDiffBackend

ForwardDiffBackend

ReverseDiffBackend

value_and_derivative

value_and_gradient

value_and_gradient!

value_and_jacobian

value_and_jacobian!

value_and_multiderivative

value_and_multiderivative!

value_and_pullback

value_and_pullback!

value_and_pushforward

value_and_pushforward!

Backends

DifferentiationInterface.ChainRulesForwardBackend — Type

ChainRulesForwardBackend{RC}

Performs autodiff with forward-mode AD packages based on ChainRulesCore.jl, like Diffractor.jl.

This must be constructed with an appropriate RuleConfig instance.

using Diffractor, DifferentiationInterface
backend = ChainRulesForwardBackend(Diffractor.DiffractorRuleConfig())

DifferentiationInterface.ChainRulesReverseBackend — Type

ChainRulesReverseBackend{RC}

Performs autodiff with reverse-mode AD packages based on ChainRulesCore.jl, like Zygote.jl.

This must be constructed with an appropriate RuleConfig instance:

using Zygote, DifferentiationInterface
backend = ChainRulesReverseBackend(Zygote.ZygoteRuleConfig())

DifferentiationInterface.EnzymeForwardBackend — Type

EnzymeForwardBackend

Performs forward-mode autodiff with Enzyme.jl.

DifferentiationInterface.EnzymeReverseBackend — Type

EnzymeReverseBackend

Performs reverse-mode autodiff with Enzyme.jl.

DifferentiationInterface.FiniteDiffBackend — Type

FiniteDiffBackend

Performs autodiff with FiniteDiff.jl.

DifferentiationInterface.ForwardDiffBackend — Type

ForwardDiffBackend

Performs autodiff with ForwardDiff.jl.

DifferentiationInterface.ReverseDiffBackend — Type

ReverseDiffBackend

Performs autodiff with ReverseDiff.jl.

Primitives

Pushforward

DifferentiationInterface.value_and_pushforward! — Method

value_and_pushforward!(dy, backend, f, x, dx[, stuff]) -> (y, dy)

Compute the Jacobian-vector product of the Jacobian of f at x with dx inside dy. Returns the primal output f(x) and the JVP dy.

Interface requirement

This is the only required implementation for a forward mode backend.

Arguments

y: primal output

dy: cotangent, might be modified

backend: forward-mode autodiff backend

f: function x -> y to differentiate

x: argument

dx: tangent

stuff: optional backend-specific storage (cache, config), might be modified

DifferentiationInterface.value_and_pushforward — Method

value_and_pushforward(backend, f, x, dx[, stuff]) -> (y, dy)

Call value_and_pushforward! after allocating memory for the Jacobian-vector product.

Pullback

DifferentiationInterface.value_and_pullback! — Method

value_and_pullback!(dx, backend, f, x, dy[, stuff]) -> (y, dx)

Compute the vector-Jacobian product of dy with the Jacobian of f at x inside dx. Returns the primal output f(x) and the VJP dx.

Interface requirement

This is the only required implementation for a reverse mode backend.

Arguments

y: primal output

dx: tangent, might be modified

backend: reverse-mode autodiff backend

f: function x -> y to differentiate

x: argument

dy: cotangent

stuff: optional backend-specific storage (cache, config), might be modified

DifferentiationInterface.value_and_pullback — Method

value_and_pullback(backend, f, x, dy[, stuff]) -> (y, dx)

Call value_and_pullback! after allocating memory for the vector-Jacobian product.

Special cases

Scalar to scalar

DifferentiationInterface.value_and_derivative — Function

value_and_derivative(backend, f, x) -> (y, der)

Compute the derivative of a scalar-to-scalar function. Returns the primal output f(x) and the derivative.

Scalar to array

DifferentiationInterface.value_and_multiderivative! — Function

value_and_multiderivative!(multider, backend, f, x) -> (y, multider)

Compute the derivative of a scalar-to-array function inside multider. Returns the primal output f(x) and the derivative.

DifferentiationInterface.value_and_multiderivative — Method

value_and_multiderivative(backend, f, x) -> (y, multider)

Call value_and_multiderivative! after allocating memory for the multiderivative.

Array to scalar

DifferentiationInterface.value_and_gradient! — Function

value_and_gradient!(grad, backend, f, x) -> (y, grad)

Compute the gradient of an array-to-scalar function inside dx. Returns the primal output f(x) and the gradient.

DifferentiationInterface.value_and_gradient — Method

value_and_gradient(backend, f, x) -> (y, grad)

Call value_and_gradient! after allocating memory for the gradient.

Array to array

DifferentiationInterface.value_and_jacobian! — Method

value_and_jacobian!(jac, backend, f, x[, stuff]) -> (y, jac)

Compute the Jacobian inside jac. Returns the primal output f(x) and the Jacobian.

Notes

For a function f: ℝⁿ → ℝᵐ, jac is expected to be a m × n matrix. If the input or output is a higher-order array, it is flattened with vec.

DifferentiationInterface.value_and_jacobian — Method

value_and_jacobian(backend, f, x) -> (y, jac)

Call value_and_jacobian! after allocating memory for the Jacobian matrix.

Internals

DifferentiationInterface.basisarray — Method

basisarray(backend, a::AbstractArray, i)

Construct the i-th stardard basis array in the vector space of a with element type eltype(a).

Note

If an AD backend benefits from a more specialized unit vector implementation, this function can be extended on the backend type.