API reference

DifferentiationInterface.DifferentiationInterface — Module

DifferentiationInterface

An interface to various automatic differentiation backends in Julia.

Exports

AutoFastDifferentiation

SecondOrder

check_available

check_hessian

check_mutation

derivative

derivative!!

gradient

gradient!!

hessian

hvp

jacobian

jacobian!!

prepare_derivative

prepare_gradient

prepare_hessian

prepare_hvp

prepare_jacobian

prepare_pullback

prepare_pushforward

prepare_second_derivative

pullback

pullback!!

pushforward

pushforward!!

second_derivative

value_and_derivative

value_and_derivative!!

value_and_gradient

value_and_gradient!!

value_and_jacobian

value_and_jacobian!!

value_and_pullback

value_and_pullback!!

value_and_pushforward

value_and_pushforward!!

Derivative

DifferentiationInterface.derivative!! — Method

derivative!!(f, der, backend, x, [extras]) -> der

DifferentiationInterface.derivative — Method

derivative(f, backend, x, [extras]) -> der

DifferentiationInterface.value_and_derivative!! — Method

value_and_derivative!!(f, der, backend, x, [extras]) -> (y, der)
value_and_derivative!!(f!, y, der, backend, x, [extras]) -> (y, der)

DifferentiationInterface.value_and_derivative — Method

value_and_derivative(f, backend, x, [extras]) -> (y, der)

Gradient

DifferentiationInterface.gradient!! — Method

gradient!!(f, grad, backend, x, [extras]) -> grad

DifferentiationInterface.gradient — Method

gradient(f, backend, x, [extras]) -> grad

DifferentiationInterface.value_and_gradient!! — Method

value_and_gradient!!(f, grad, backend, x, [extras]) -> (y, grad)

DifferentiationInterface.value_and_gradient — Method

value_and_gradient(f, backend, x, [extras]) -> (y, grad)

Jacobian

DifferentiationInterface.jacobian!! — Method

jacobian!!(f, jac, backend, x, [extras]) -> jac

DifferentiationInterface.jacobian — Method

jacobian(f, backend, x, [extras]) -> jac

DifferentiationInterface.value_and_jacobian!! — Method

value_and_jacobian!!(f, jac, backend, x, [extras]) -> (y, jac)
value_and_jacobian!!(f!, y, jac, backend, x, [extras]) -> (y, jac)

DifferentiationInterface.value_and_jacobian — Method

value_and_jacobian(f, backend, x, [extras]) -> (y, jac)

Second order

DifferentiationInterface.SecondOrder — Type

SecondOrder

Combination of two backends for second-order differentiation.

Fields

outer::ADTypes.AbstractADType: backend for the outer differentiation

inner::ADTypes.AbstractADType: backend for the inner differentiation

DifferentiationInterface.second_derivative — Method

second_derivative(f, backend, x, [extras]) -> der2

DifferentiationInterface.hessian — Method

hessian(f, backend, x, [extras]) -> hess

DifferentiationInterface.hvp — Method

hvp(f, backend, x, v, [extras]) -> p

Primitives

DifferentiationInterface.pushforward!! — Method

pushforward!!(f, dy, backend, x, dx, [extras]) -> (y, dy)

DifferentiationInterface.pushforward — Method

pushforward(f, backend, x, dx, [extras]) -> (y, dy)

DifferentiationInterface.value_and_pushforward!! — Method

value_and_pushforward!!(f, dy, backend, x, dx, [extras]) -> (y, dy)
value_and_pushforward!!(f!, y, dy, backend, x, dx, [extras]) -> (y, dy)

DifferentiationInterface.value_and_pushforward — Method

value_and_pushforward(f, backend, x, dx, [extras]) -> (y, dy)

DifferentiationInterface.pullback!! — Method

pullback!!(f, dx, backend, x, dy, [extras]) -> dx

DifferentiationInterface.pullback — Method

pullback(f, backend, x, dy, [extras]) -> dx

DifferentiationInterface.value_and_pullback!! — Method

value_and_pullback!!(f, dx, backend, x, dy, [extras]) -> (y, dx)
value_and_pullback!!(f!, y, dx, backend, x, dy, [extras]) -> (y, dx)

DifferentiationInterface.value_and_pullback — Method

value_and_pullback(f, backend, x, dy, [extras]) -> (y, dx)

Backend queries

DifferentiationInterface.check_available — Method

check_available(backend)

Check whether backend is available by trying a scalar-to-scalar derivative.

Warning

Might take a while due to compilation time.

DifferentiationInterface.check_hessian — Method

check_hessian(backend)

Check whether backend supports second order differentiation by trying a hessian.

Warning

Might take a while due to compilation time.

DifferentiationInterface.check_mutation — Method

check_mutation(backend)

Check whether backend supports differentiation of mutating functions by trying a jacobian.

Warning

Might take a while due to compilation time.

Preparation

DifferentiationInterface.prepare_derivative — Method

prepare_derivative(f, backend, x) -> extras
prepare_derivative(f!, backend, y, x) -> extras

Create an extras object that can be given to derivative operators.

DifferentiationInterface.prepare_gradient — Method

prepare_gradient(f, backend, x) -> extras

Create an extras object that can be given to gradient operators.

DifferentiationInterface.prepare_hessian — Method

prepare_hessian(f, backend, x) -> extras

Create an extras object that can be given to Hessian operators.

DifferentiationInterface.prepare_hvp — Method

prepare_hvp(f, backend, x) -> extras

Create an extras object that can be given to Hessian-vector product operators.

DifferentiationInterface.prepare_jacobian — Method

prepare_jacobian(f, backend, x) -> extras
prepare_jacobian(f!, backend, y, x) -> extras

Create an extras object that can be given to Jacobian operators.

DifferentiationInterface.prepare_pullback — Method

prepare_pullback(f, backend, x) -> extras
prepare_pullback(f!, backend, y, x) -> extras

Create an extras object that can be given to pullback operators.

DifferentiationInterface.prepare_pushforward — Method

prepare_pushforward(f, backend, x) -> extras
prepare_pushforward(f!, backend, y, x) -> extras

Create an extras object that can be given to pushforward operators.

DifferentiationInterface.prepare_second_derivative — Method

prepare_second_derivative(f, backend, x) -> extras
prepare_second_derivative(f!, backend, y, x) -> extras

Create an extras object that can be given to second derivative operators.

Testing & benchmarking

DifferentiationInterface.DifferentiationTest — Module

DifferentiationInterface.DifferentiationTest

Testing utilities for DifferentiationInterface.

DifferentiationInterface.DifferentiationTest.BenchmarkData — Type

BenchmarkData

Ad-hoc storage type for differentiation benchmarking results. You can turn it into a DataFrame as follows:

df = DataFrames.DataFrame(pairs(benchmark_data)...)

Fields

These are not part of the public API.

backend::Vector{String}

mode::Vector{Type}

operator::Vector{Function}

variant::Vector{Function}

func::Vector{String}

mutating::Vector{Bool}

input_type::Vector{Type}

output_type::Vector{Type}

input_size::Vector

output_size::Vector

samples::Vector{Int64}

time::Vector{Float64}

bytes::Vector{Float64}

allocs::Vector{Float64}

compile_fraction::Vector{Float64}

gc_fraction::Vector{Float64}

evals::Vector{Float64}

DifferentiationInterface.DifferentiationTest.Scenario — Type

Scenario{mutating}

Store a testing scenario composed of a function and its input + output + tangents.

Fields

f::Any: function

x::Any: input

y::Any: output

dx::Any: pushforward seed

dy::Any: pullback seed

DifferentiationInterface.DifferentiationTest.default_scenarios — Method

default_scenarios()

Create a vector of Scenarios for testing differentiation.

DifferentiationInterface.DifferentiationTest.test_operators — Function

test_operators(backends, [operators, scenarios]; [kwargs...])

Cross-test a list of backends for a list of operators on a list of scenarios, running a variety of different tests.

If benchmark is false, this returns a TestSet object.

If benchmark is true, this returns a BenchmarkData object, which is easy to turn into a DataFrame.

Default arguments

operators::Vector{Function}: the list [pushforward, pullback,derivative, gradient, jacobian, second_derivative, hvp, hessian]

scenarios::Vector{Scenario}: the output of default_scenarios()

Keyword arguments

Testing:

correctness=true: whether to compare the differentiation results with those given by ForwardDiff.jl

type_stability=true: whether to check type stability with JET.jl

call_count=false: whether to check that the function is called the right number of times

benchmark=false: whether to run and return a benchmark suite with Chairmarks.jl

allocations=false: whether to check that the benchmarks are allocation-free

detailed=false: whether to print a detailed test set (by scenario) or condensed test set (by operator)

Filtering:

input_type=Any: restrict scenario inputs to subtypes of this

output_type=Any: restrict scenario outputs to subtypes of this

first_order=true: consider first order operators

second_order=true: consider second order operators

allocating=true: consider operators for allocating functions

mutating=true: consider operators for mutating functions

excluded=Symbol[]: list of excluded operators

DifferentiationInterface.DifferentiationTest.test_operators — Method

test_operators(
    backend::ADTypes.AbstractADType,
    args...;
    kwargs...
) -> Union{Test.FallbackTestSet, BenchmarkData, Test.DefaultTestSet}

Shortcut for a single backend.

Internals

This is not part of the public API.

DifferentiationInterface.basisarray — Method

basisarray(backend, a::AbstractArray, i::CartesianIndex)

Construct the i-th stardard basis array in the vector space of a with element type eltype(a).

Note

If an AD backend benefits from a more specialized basis array implementation, this function can be extended on the backend type.

DifferentiationInterface.mode — Method

mode(backend)

Return the AD mode of a backend in a statically predictable way.

The return value is a Type object chosen among:

ADTypes.AbstractForwardMode

ADTypes.AbstractFiniteDifferencesMode

ADTypes.AbstractReverseMode

ADTypes.AbstractSymbolicDifferentiationMode

This function exists because there are backends (like Enzyme) that can support both forward and reverse mode, which means their ADTypes.jl object does not subtype either.

DifferentiationInterface.supports_mutation — Method

supports_mutation(backend)

Return MutationSupported or MutationNotSupported in a statically predictable way.

DifferentiationInterface.supports_pullback — Method

supports_pullback(backend)

Return PullbackSupported or PullbackNotSupported in a statically predictable way.

DifferentiationInterface.supports_pushforward — Method

supports_pushforward(backend)

Return PushforwardSupported or PushforwardNotSupported in a statically predictable way.

DifferentiationInterface.ForwardOverForward — Type

ForwardOverForward

Traits identifying second-order backends that compute HVPs in forward over forward mode (inefficient).

DifferentiationInterface.ForwardOverReverse — Type

ForwardOverReverse

Traits identifying second-order backends that compute HVPs in forward over reverse mode.

DifferentiationInterface.MutationNotSupported — Type

MutationNotSupported

Trait identifying backends that do not support mutating functions f!(y, x).

DifferentiationInterface.MutationSupported — Type

MutationSupported

Trait identifying backends that support mutating functions f!(y, x).

DifferentiationInterface.PullbackNotSupported — Type

PullbackNotSupported

Trait identifying backends that do not support efficient pullbacks.

DifferentiationInterface.PullbackSupported — Type

PullbackSupported

Trait identifying backends that support efficient pullbacks.

DifferentiationInterface.PushforwardNotSupported — Type

PushforwardNotSupported

Trait identifying backends that do not support efficient pushforwards.

DifferentiationInterface.PushforwardSupported — Type

PushforwardSupported

Trait identifying backends that support efficient pushforwards.

DifferentiationInterface.ReverseOverForward — Type

ReverseOverForward

Traits identifying second-order backends that compute HVPs in reverse over forward mode.

DifferentiationInterface.ReverseOverReverse — Type

ReverseOverReverse

Traits identifying second-order backends that compute HVPs in reverse over reverse mode.

DifferentiationInterface.DifferentiationTest.AutoZeroForward — Type

AutoZeroForward <: ADTypes.AbstractForwardMode

Trivial backend that sets all derivatives to zero. Used in testing and benchmarking.

DifferentiationInterface.DifferentiationTest.AutoZeroReverse — Type

AutoZeroReverse <: ADTypes.AbstractReverseMode

Trivial backend that sets all derivatives to zero. Used in testing and benchmarking.

DifferentiationInterface.DifferentiationTest.backend_string — Method

backend_string(backend)

Return a shorter string than the full object printing from ADTypes.jl. Might be ambiguous.