API reference

DifferentiationInterface.DifferentiationInterface — Module

DifferentiationInterface

An interface to various automatic differentiation backends in Julia.

Exports

derivative

gradient

gradient!

jacobian

jacobian!

multiderivative

multiderivative!

prepare_derivative

prepare_gradient

prepare_jacobian

prepare_multiderivative

prepare_pullback

prepare_pushforward

pullback

pullback!

pushforward

pushforward!

value_and_derivative

value_and_gradient

value_and_gradient!

value_and_jacobian

value_and_jacobian!

value_and_multiderivative

value_and_multiderivative!

value_and_pullback

value_and_pullback!

value_and_pushforward

value_and_pushforward!

Derivative

DifferentiationInterface.derivative — Method

derivative(backend, f, x, [extras]) -> der

Compute the derivative der = f'(x) of a scalar-to-scalar function.

DifferentiationInterface.value_and_derivative — Method

value_and_derivative(backend, f, x, [extras]) -> (y, der)

Compute the primal value y = f(x) and the derivative der = f'(x) of a scalar-to-scalar function.

Multiderivative

DifferentiationInterface.multiderivative! — Method

multiderivative!(multider, backend, f, x, [extras]) -> multider

Compute the (array-valued) derivative multider = f'(x) of a scalar-to-array function, overwriting multider if possible.

DifferentiationInterface.multiderivative — Method

multiderivative(backend, f, x, [extras]) -> multider

Compute the (array-valued) derivative multider = f'(x) of a scalar-to-array function.

DifferentiationInterface.value_and_multiderivative! — Method

value_and_multiderivative!(multider, backend, f, x, [extras]) -> (y, multider)

Compute the primal value y = f(x) and the (array-valued) derivative multider = f'(x) of a scalar-to-array function, overwriting multider if possible.

DifferentiationInterface.value_and_multiderivative — Method

value_and_multiderivative(backend, f, x, [extras]) -> (y, multider)

Compute the primal value y = f(x) and the (array-valued) derivative multider = f'(x) of a scalar-to-array function.

Gradient

DifferentiationInterface.gradient! — Method

gradient!(grad, backend, f, x, [extras]) -> grad

Compute the gradient grad = ∇f(x) of an array-to-scalar function, overwriting grad if possible.

DifferentiationInterface.gradient — Method

gradient(backend, f, x, [extras]) -> grad

Compute the gradient grad = ∇f(x) of an array-to-scalar function.

DifferentiationInterface.value_and_gradient! — Method

value_and_gradient!(grad, backend, f, x, [extras]) -> (y, grad)

Compute the primal value y = f(x) and the gradient grad = ∇f(x) of an array-to-scalar function, overwriting grad if possible.

DifferentiationInterface.value_and_gradient — Method

value_and_gradient(backend, f, x, [extras]) -> (y, grad)

Compute the primal value y = f(x) and the gradient grad = ∇f(x) of an array-to-scalar function.

Jacobian

DifferentiationInterface.jacobian! — Method

jacobian!(jac, backend, f, x, [extras]) -> jac

Compute the Jacobian matrix jac = ∂f(x) of an array-to-array function, overwriting jac if possible.

Notes

Regardless of the shape of x and y, if x has length n and y has length m, then jac is expected to be a m × n matrix. This function acts as if the input and output had been flattened with vec.

DifferentiationInterface.jacobian — Method

jacobian(backend, f, x, [extras]) -> jac

Compute the Jacobian matrix jac = ∂f(x) of an array-to-array function.

Notes

Regardless of the shape of x and y, if x has length n and y has length m, then jac is expected to be a m × n matrix. This function acts as if the input and output had been flattened with vec.

DifferentiationInterface.value_and_jacobian! — Method

value_and_jacobian!(jac, backend, f, x, [extras]) -> (y, jac)

Compute the primal value y = f(x) and the Jacobian matrix jac = ∂f(x) of an array-to-array function, overwriting jac if possible.

Notes

Regardless of the shape of x and y, if x has length n and y has length m, then jac is expected to be a m × n matrix. This function acts as if the input and output had been flattened with vec.

DifferentiationInterface.value_and_jacobian — Method

value_and_jacobian(backend, f, x, [extras]) -> (y, jac)

Compute the primal value y = f(x) and the Jacobian matrix jac = ∂f(x) of an array-to-array function.

Notes

Regardless of the shape of x and y, if x has length n and y has length m, then jac is expected to be a m × n matrix. This function acts as if the input and output had been flattened with vec.

Pushforward (JVP)

DifferentiationInterface.pushforward — Function

pushforward(backend, f, x, dx, [extras]) -> dy

Compute the Jacobian-vector product dy = ∂f(x) * dx.

DifferentiationInterface.pushforward! — Function

pushforward!(dy, backend, f, x, dx, [extras]) -> dy

Compute the Jacobian-vector product dy = ∂f(x) * dx, overwriting dy if possible.

DifferentiationInterface.value_and_pushforward — Function

value_and_pushforward(backend, f, x, dx, [extras]) -> (y, dy)

Compute the primal value y = f(x) and the Jacobian-vector product dy = ∂f(x) * dx.

DifferentiationInterface.value_and_pushforward! — Function

value_and_pushforward!(dy, backend, f, x, dx, [extras]) -> (y, dy)

Compute the primal value y = f(x) and the Jacobian-vector product dy = ∂f(x) * dx, overwriting dy if possible.

Interface requirement

This is the only required implementation for a forward mode backend.

Pullback (JVP)

DifferentiationInterface.pullback — Function

pullback(backend, f, x, dy, [extras]) -> dx

Compute the vector-Jacobian product dx = ∂f(x)' * dy.

DifferentiationInterface.pullback! — Function

pullback!(dx, backend, f, x, dy, [extras]) -> dx

Compute the vector-Jacobian product dx = ∂f(x)' * dy, overwriting dx if possible.

DifferentiationInterface.value_and_pullback — Function

value_and_pullback(backend, f, x, dy, [extras]) -> (y, dx)

Compute the primal value y = f(x) and the vector-Jacobian product dx = ∂f(x)' * dy.

DifferentiationInterface.value_and_pullback! — Function

value_and_pullback!(dx, backend, f, x, dy, [extras]) -> (y, dx)

Compute the primal value y = f(x) and the vector-Jacobian product dx = ∂f(x)' * dy, overwriting dx if possible.

Interface requirement

This is the only required implementation for a reverse mode backend.

Preparation

DifferentiationInterface.prepare_derivative — Method

prepare_derivative(backend, f, x) -> extras

Create an extras object that can be given to derivative operators.

DifferentiationInterface.prepare_gradient — Method

prepare_gradient(backend, f, x) -> extras

Create an extras object that can be given to gradient operators.

DifferentiationInterface.prepare_jacobian — Method

prepare_jacobian(backend, f, x) -> extras

Create an extras object that can be given to jacobian operators.

DifferentiationInterface.prepare_multiderivative — Method

prepare_multiderivative(backend, f, x) -> extras

Create an extras object that can be given to multiderivative operators.

DifferentiationInterface.prepare_pullback — Method

prepare_pullback(backend, f, x) -> extras

Create an extras object that can be given to pullback operators.

DifferentiationInterface.prepare_pushforward — Method

prepare_pushforward(backend, f, x) -> extras

Create an extras object that can be given to pushforward operators.

Internals

These are not part of the public API.

DifferentiationInterface.autodiff_mode — Method

autodiff_mode(backend)

Return ForwardMode() or ReverseMode() in a statically predictable way.

This function must be overloaded for backends that do not inherit from ADTypes.AbstractForwardMode or ADTypes.AbstractReverseMode (e.g. because they support both forward and reverse).

We classify ADTypes.AbstractFiniteDifferencesMode as forward mode.

DifferentiationInterface.ForwardMode — Type

ForwardMode

Trait identifying forward mode AD backends. Used for internal dispatch only.

DifferentiationInterface.ReverseMode — Type

ReverseMode

Trait identifying reverse mode AD backends. Used for internal dispatch only.

DifferentiationInterface.basisarray — Method

basisarray(backend, a::AbstractArray, i::CartesianIndex)

Construct the i-th stardard basis array in the vector space of a with element type eltype(a).

Note

If an AD backend benefits from a more specialized basis array implementation, this function can be extended on the backend type.