API reference

DifferentiationInterface.DifferentiationInterface — Module

DifferentiationInterface

An interface to various automatic differentiation backends in Julia.

Exports

AutoFastDifferentiation

SecondOrder

derivative

gradient

gradient!

gradient_and_hessian_vector_product

gradient_and_hessian_vector_product!

hessian

hessian!

hessian_vector_product

hessian_vector_product!

jacobian

jacobian!

multiderivative

multiderivative!

prepare_derivative

prepare_gradient

prepare_hessian

prepare_hessian_vector_product

prepare_jacobian

prepare_multiderivative

prepare_pullback

prepare_pushforward

prepare_second_derivative

pullback

pullback!

pushforward

pushforward!

second_derivative

value_and_derivative

value_and_gradient

value_and_gradient!

value_and_jacobian

value_and_jacobian!

value_and_multiderivative

value_and_multiderivative!

value_and_pullback

value_and_pullback!

value_and_pushforward

value_and_pushforward!

value_derivative_and_second_derivative

value_gradient_and_hessian

value_gradient_and_hessian!

Derivative

DifferentiationInterface.derivative — Function

derivative(backend, f, x, [extras]) -> der

Compute the derivative der = f'(x) of a scalar-to-scalar function.

DifferentiationInterface.value_and_derivative — Function

value_and_derivative(backend, f, x, [extras]) -> (y, der)

Compute the primal value y = f(x) and the derivative der = f'(x) of a scalar-to-scalar function.

Multiderivative

DifferentiationInterface.multiderivative — Function

multiderivative(backend, f, x, [extras]) -> multider

Compute the (array-valued) derivative multider = f'(x) of a scalar-to-array function.

DifferentiationInterface.multiderivative! — Function

multiderivative!(multider, backend, f, x, [extras]) -> multider

Compute the (array-valued) derivative multider = f'(x) of a scalar-to-array function, overwriting multider.

DifferentiationInterface.value_and_multiderivative — Function

value_and_multiderivative(backend, f, x, [extras]) -> (y, multider)

Compute the primal value y = f(x) and the (array-valued) derivative multider = f'(x) of a scalar-to-array function.

DifferentiationInterface.value_and_multiderivative! — Function

value_and_multiderivative!(multider, backend, f, x, [extras]) -> (y, multider)
value_and_multiderivative!(y, multider, backend, f!, x, [extras]) -> (y, multider)

Compute the primal value y = f(x) and the (array-valued) derivative multider = f'(x) of a scalar-to-array function, overwriting multider.

Gradient

DifferentiationInterface.gradient — Function

gradient(backend, f, x, [extras]) -> grad

Compute the gradient grad = ∇f(x) of an array-to-scalar function.

DifferentiationInterface.gradient! — Function

gradient!(grad, backend, f, x, [extras]) -> grad

Compute the gradient grad = ∇f(x) of an array-to-scalar function, overwriting grad.

DifferentiationInterface.value_and_gradient — Function

value_and_gradient(backend, f, x, [extras]) -> (y, grad)

Compute the primal value y = f(x) and the gradient grad = ∇f(x) of an array-to-scalar function.

DifferentiationInterface.value_and_gradient! — Function

value_and_gradient!(grad, backend, f, x, [extras]) -> (y, grad)

Compute the primal value y = f(x) and the gradient grad = ∇f(x) of an array-to-scalar function, overwriting grad.

Jacobian

DifferentiationInterface.jacobian — Function

jacobian(backend, f, x, [extras]) -> jac

Compute the Jacobian matrix jac = ∂f(x) of an array-to-array function.

Notes

Regardless of the shape of x and y, if x has length n and y has length m, then jac is expected to be a m × n matrix. This function acts as if the input and output had been flattened with vec.

DifferentiationInterface.jacobian! — Function

jacobian!(jac, backend, f, x, [extras]) -> jac

Compute the Jacobian matrix jac = ∂f(x) of an array-to-array function, overwriting jac.

Notes

Regardless of the shape of x and y, if x has length n and y has length m, then jac is expected to be a m × n matrix. This function acts as if the input and output had been flattened with vec.

DifferentiationInterface.value_and_jacobian — Function

value_and_jacobian(backend, f, x, [extras]) -> (y, jac)

Compute the primal value y = f(x) and the Jacobian matrix jac = ∂f(x) of an array-to-array function.

Notes

Regardless of the shape of x and y, if x has length n and y has length m, then jac is expected to be a m × n matrix. This function acts as if the input and output had been flattened with vec.

DifferentiationInterface.value_and_jacobian! — Function

value_and_jacobian!(jac, backend, f, x, [extras]) -> (y, jac)
value_and_jacobian!(y, jac, backend, f!, x, [extras]) -> (y, jac)

Compute the primal value y = f(x) and the Jacobian matrix jac = ∂f(x) of an array-to-array function, overwriting jac.

Notes

Regardless of the shape of x and y, if x has length n and y has length m, then jac is expected to be a m × n matrix. This function acts as if the input and output had been flattened with vec.

Second derivative

DifferentiationInterface.second_derivative — Function

second_derivative(backend, f, x, [extras]) -> derder

Compute the second derivative derder = f''(x) of a scalar-to-scalar function.

DifferentiationInterface.value_derivative_and_second_derivative — Function

value_derivative_and_second_derivative(backend, f, x, [extras]) -> (y, der, derder)

Compute the primal value y = f(x), the derivative der = f'(x) and the second derivative derder = f''(x) of a scalar-to-scalar function.

Hessian

DifferentiationInterface.hessian — Function

hessian(backend, f, x, [extras]) -> hess

Compute the Hessian hess = ∇²f(x) of an array-to-scalar function.

Notes

Regardless of the shape of x, if x has length n, then hess is expected to be a n × n matrix. This function acts as if the input had been flattened with vec.

DifferentiationInterface.hessian! — Function

hessian!(hess, backend, f, x, [extras]) -> hess

Compute the Hessian hess = ∇²f(x) of an array-to-scalar function, overwriting hess.

Notes

Regardless of the shape of x, if x has length n, then hess is expected to be a n × n matrix. This function acts as if the input had been flattened with vec.

DifferentiationInterface.value_gradient_and_hessian — Function

value_gradient_and_hessian(backend, f, x, [extras]) -> (y, grad, hess)

Compute the primal value y = f(x), the gradient grad = ∇f(x) and the Hessian hess = ∇²f(x) of an array-to-scalar function, overwriting grad and hess.

Notes

Regardless of the shape of x, if x has length n, then hess is expected to be a n × n matrix. This function acts as if the input had been flattened with vec.

DifferentiationInterface.value_gradient_and_hessian! — Function

value_gradient_and_hessian!(grad, hess, backend, f, x, [extras]) -> (y, grad, hess)

Compute the primal value y = f(x), the gradient grad = ∇f(x) and the Hessian hess = ∇²f(x) of an array-to-scalar function, overwriting grad and hess.

Notes

Regardless of the shape of x, if x has length n, then hess is expected to be a n × n matrix. This function acts as if the input had been flattened with vec.

Pushforward (JVP)

DifferentiationInterface.pushforward — Function

pushforward(backend, f, x, dx, [extras]) -> dy

Compute the Jacobian-vector product dy = ∂f(x) * dx.

DifferentiationInterface.pushforward! — Function

pushforward!(dy, backend, f, x, dx, [extras]) -> dy

Compute the Jacobian-vector product dy = ∂f(x) * dx, overwriting dy if possible.

DifferentiationInterface.value_and_pushforward — Function

value_and_pushforward(backend, f, x, dx, [extras]) -> (y, dy)

Compute the primal value y = f(x) and the Jacobian-vector product dy = ∂f(x) * dx.

DifferentiationInterface.value_and_pushforward! — Method

value_and_pushforward!(dy, backend, f, x, dx, [extras]) -> (y, dy)
value_and_pushforward!(y, dy, backend, f!, x, dx, [extras]) -> (y, dy)

Compute the primal value y = f(x) and the Jacobian-vector product dy = ∂f(x) * dx, overwriting dy if possible.

Interface requirement

This is the only required implementation for a forward mode backend.

Pullback (JVP)

DifferentiationInterface.pullback — Function

pullback(backend, f, x, dy, [extras]) -> dx

Compute the vector-Jacobian product dx = ∂f(x)' * dy.

DifferentiationInterface.pullback! — Function

pullback!(dx, backend, f, x, dy, [extras]) -> dx

Compute the vector-Jacobian product dx = ∂f(x)' * dy, overwriting dx if possible.

DifferentiationInterface.value_and_pullback — Function

value_and_pullback(backend, f, x, dy, [extras]) -> (y, dx)

Compute the primal value y = f(x) and the vector-Jacobian product dx = ∂f(x)' * dy.

DifferentiationInterface.value_and_pullback! — Method

value_and_pullback!(dx, backend, f, x, dy, [extras]) -> (y, dx)
value_and_pullback!(y, dx, backend, f!, x, dy, [extras]) -> (y, dx)

Compute the primal value y = f(x) and the vector-Jacobian product dx = ∂f(x)' * dy, overwriting dx if possible.

Interface requirement

This is the only required implementation for a reverse mode backend.

Hessian-vector product (HVP)

DifferentiationInterface.gradient_and_hessian_vector_product — Function

gradient_and_hessian_vector_product(backend, f, x, v, [extras]) -> (grad, hvp)

Compute the gradient grad = ∇f(x) and the Hessian-vector product hvp = ∇²f(x) * v of an array-to-scalar function.

Warning

Only works with a forward outer mode.

DifferentiationInterface.gradient_and_hessian_vector_product! — Function

gradient_and_hessian_vector_product!(grad, backend, hvp, backend, f, x, v, [extras]) -> (grad, hvp)

Compute the gradient grad = ∇f(x) and the Hessian-vector product hvp = ∇²f(x) * v of an array-to-scalar function, overwriting grad and hvp.

Warning

Only works with a forward outer mode.

DifferentiationInterface.hessian_vector_product — Function

hessian_vector_product(backend, f, x, v, [extras]) -> hvp

Compute the Hessian-vector product hvp = ∇²f(x) * v of an array-to-scalar function.

DifferentiationInterface.hessian_vector_product! — Function

hessian_vector_product!(hvp, backend, f, x, v, [extras]) -> hvp

Compute the Hessian-vector product hvp = ∇²f(x) * v of an array-to-scalar function, overwriting hvp.

Preparation

DifferentiationInterface.prepare_derivative — Method

prepare_derivative(backend, f, x) -> extras

Create an extras object that can be given to derivative operators.

DifferentiationInterface.prepare_gradient — Method

prepare_gradient(backend, f, x) -> extras

Create an extras object that can be given to gradient operators.

DifferentiationInterface.prepare_hessian — Method

prepare_hessian(backend, f, x) -> extras

Create an extras object that can be given to Hessian operators.

DifferentiationInterface.prepare_hessian_vector_product — Method

prepare_hessian_vector_product(backend, f, x) -> extras

Create an extras object that can be given to Hessian-vector product operators.

DifferentiationInterface.prepare_jacobian — Method

prepare_jacobian(backend, f, x) -> extras
prepare_jacobian(backend, f!, x, y) -> extras

Create an extras object that can be given to Jacobian operators.

DifferentiationInterface.prepare_multiderivative — Method

prepare_multiderivative(backend, f, x) -> extras
prepare_multiderivative(backend, f!, x, y) -> extras

Create an extras object that can be given to multiderivative operators.

DifferentiationInterface.prepare_pullback — Method

prepare_pullback(backend, f, x) -> extras
prepare_pullback(backend, f!, x, y) -> extras

Create an extras object that can be given to pullback operators.

DifferentiationInterface.prepare_pushforward — Method

prepare_pushforward(backend, f, x) -> extras
prepare_pushforward(backend, f!, x, y) -> extras

Create an extras object that can be given to pushforward operators.

DifferentiationInterface.prepare_second_derivative — Method

prepare_second_derivative(backend, f, x) -> extras

Create an extras object that can be given to second derivative operators.

Internals

These are not part of the public API.

DifferentiationInterface.AutoZeroForward — Type

AutoZeroForward <: ADTypes.AbstractForwardMode

Trivial backend that sets all derivatives to zero. Used in testing and benchmarking.

DifferentiationInterface.AutoZeroReverse — Type

AutoZeroReverse <: ADTypes.AbstractReverseMode

Trivial backend that sets all derivatives to zero. Used in testing and benchmarking.

DifferentiationInterface.ForwardMode — Type

ForwardMode

Trait identifying forward mode (and finite differences) first-order AD backends.

DifferentiationInterface.ReverseMode — Type

ReverseMode

Trait identifying reverse mode first-order AD backends.

DifferentiationInterface.SymbolicMode — Type

SymbolicMode

Trait identifying symbolic first-order AD backends. Their fallback structure is different from the rest.

DifferentiationInterface.basisarray — Method

basisarray(backend, a::AbstractArray, i::CartesianIndex)

Construct the i-th stardard basis array in the vector space of a with element type eltype(a).

Note

If an AD backend benefits from a more specialized basis array implementation, this function can be extended on the backend type.

DifferentiationInterface.mode — Function

mode(backend)

Return the AD mode of a backend in a statically predictable way.

This function must be overloaded for backends that support both forward and reverse.

We classify finite differences as a forward mode.

Submodules

These are not part of the public API.

DifferentiationInterface.DifferentiationTest — Module

DifferentiationInterface.DifferentiationTest

Testing utilities for DifferentiationInterface.

DifferentiationInterface.DifferentiationTest.Scenario — Type

Scenario

Store a testing scenario composed of a function and its input + output + tangents.

Fields

f::Any: function

mutating::Bool: mutation

x::Union{Number, AbstractArray}: input

y::Union{Number, AbstractArray}: output

dx::Union{Number, AbstractArray}: pushforward seed

dy::Union{Number, AbstractArray}: pullback seed

DifferentiationInterface.DifferentiationTest.test_operators — Function

test_operators(backend::ADTypes.AbstractADType; ...)
test_operators(
    backend::ADTypes.AbstractADType,
    scenarios::Vector{<:Scenario};
    correctness,
    type_stability,
    allocations,
    input_type,
    output_type,
    first_order,
    second_order,
    allocating,
    mutating,
    operators,
    excluded
)