Backends

Types

We support all dense backend choices from ADTypes.jl, as well as their sparse wrapper AutoSparse.

For sparse backends, only the Jacobian and Hessian operators are implemented differently, the other operators behave the same as for the corresponding dense backend.

Backend	Availability	Two-argument functions	Hessian support	Example

Diffractor (forward/reverse)	✅	❌	❌	AutoDiffractor()

Enzyme (forward)	✅	✅	❌	AutoEnzyme(; mode=Enzyme.Forward)

Enzyme (reverse)	✅	✅	❌	AutoEnzyme(; mode=Enzyme.Reverse)

FastDifferentiation (symbolic)	✅	✅	✅	AutoFastDifferentiation()

FiniteDiff (forward)	✅	✅	✅	AutoFiniteDiff()

FiniteDifferences (forward)	✅	❌	✅	AutoFiniteDifferences(; fdm=FiniteDifferences.central_fdm(3, 1))

ForwardDiff (forward)	✅	✅	✅	AutoForwardDiff()

PolyesterForwardDiff (forward)	✅	✅	✅	AutoPolyesterForwardDiff(; chunksize=1)

ReverseDiff (reverse)	✅	✅	✅	AutoReverseDiff()

Symbolics (symbolic)	✅	✅	✅	AutoSymbolics()

Tapir (reverse)	✅	✅	❌	AutoTapir()

Tracker (reverse)	✅	❌	❌	AutoTracker()

Zygote (reverse)	✅	❌	✅	AutoZygote()

Compatibility with Julia 1.6

As of version 0.3.4, DifferentiationInterface.jl is compatible with Julia 1.6, the Long Term Support (LTS) version of the language. However, we were only able to test the following backends on LTS:

FiniteDifferences.jl

ForwardDiff.jl

ReverseDiff.jl

Tracker.jl

Zygote.jl

We strongly recommend that users upgrade to Julia 1.10, where all backends are tested.

Checks

Availability

You can use check_available to verify whether a given backend is loaded.

Support for two-argument functions

All backends are compatible with one-argument functions f(x) = y. Only some are compatible with two-argument functions f!(y, x) = nothing. You can check this compatibility using check_twoarg.

Support for Hessian

Only some backends are able to compute Hessians. You can use check_hessian to check this feature (beware that it will try to compute a small Hessian, so it is not instantaneous).

API reference

Warning

The following documentation has been borrowed from ADTypes.jl. Refer to the ADTypes documentation for more information.

ADTypes.ADTypes — Module

ADTypes.jl

ADTypes.jl is a multi-valued logic system to choose an automatic differentiation (AD) package and specify its parameters.

ADTypes.AbstractADType — Type

AbstractADType

Abstract supertype for all AD choices.

ADTypes.AutoChainRules — Type

AutoChainRules{RC}

Chooses any AD library based on ChainRulesCore.jl (see the list here).

Fields

ruleconfig::RC: a ChainRulesCore.RuleConfig object.

Constructor

AutoChainRules(; ruleconfig)

ADTypes.AutoDiffractor — Type

AutoDiffractor

Chooses Diffractor.jl.

Constructor

AutoDiffractor()

ADTypes.AutoEnzyme — Type

AutoEnzyme{M}

Chooses Enzyme.jl.

Fields

mode::M: can be either

an object subtyping EnzymeCore.Mode (like EnzymeCore.Forward or EnzymeCore.Reverse) if a specific mode is required

nothing to choose the best mode automatically

Constructors

AutoEnzyme(; mode=nothing)

ADTypes.AutoFastDifferentiation — Type

AutoFastDifferentiation

Chooses FastDifferentiation.jl.

Constructor

AutoFastDifferentiation()

ADTypes.AutoFiniteDiff — Type

AutoFiniteDiff{T1,T2,T3}

Chooses FiniteDiff.jl.

Fields

fdtype::T1: finite difference type

fdjtype::T2: finite difference type for the Jacobian

fdhtype::T3: finite difference type for the Hessian

Constructor

AutoFiniteDiff(; fdtype=Val(:forward), fdjtype=fdtype, fdhtype=Val(:hcentral))

ADTypes.AutoFiniteDifferences — Type

AutoFiniteDifferences{T}

Chooses FiniteDifferences.jl.

Fields

fdm::T: a FiniteDifferenceMethod

Constructor

AutoFiniteDifferences(; fdm)

ADTypes.AutoForwardDiff — Type

AutoForwardDiff{chunksize,T}

Chooses ForwardDiff.jl.

Type parameters

chunksize: the preferred chunk size to evaluate several derivatives at once

Fields

tag::T: a custom tag to handle nested differentiation calls (usually not necessary)

Constructors

AutoForwardDiff(; chunksize=nothing, tag=nothing)

ADTypes.AutoPolyesterForwardDiff — Type

AutoPolyesterForwardDiff{chunksize,T}

Chooses PolyesterForwardDiff.jl.

Type parameters

chunksize: the preferred chunk size to evaluate several derivatives at once

Fields

tag::T: a custom tag to handle nested differentiation calls (usually not necessary)

Constructors

AutoPolyesterForwardDiff(; chunksize=nothing, tag=nothing)

ADTypes.AutoReverseDiff — Type

AutoReverseDiff

Chooses ReverseDiff.jl.

Fields

compile::Bool: whether to compile the tape prior to differentiation

Constructor

AutoReverseDiff(; compile=false)

ADTypes.AutoSymbolics — Type

AutoSymbolics

Chooses Symbolics.jl.

Constructor

AutoSymbolics()

ADTypes.AutoTapir — Type

AutoTapir

Chooses Tapir.jl.

Constructor

AutoTapir()

ADTypes.AutoTracker — Type

AutoTracker

Chooses Tracker.jl.

Constructor

AutoTracker()

ADTypes.AutoZygote — Type

AutoZygote

Chooses Zygote.jl.

Constructor

AutoZygote()

ADTypes.AutoSparse — Type

AutoSparse{D,S,C}

Wraps an ADTypes.jl object to deal with sparse Jacobians and Hessians.

Fields

dense_ad::D: the underlying AD package, subtyping AbstractADType

sparsity_detector::S: the sparsity pattern detector, subtyping AbstractSparsityDetector

coloring_algorithm::C: the coloring algorithm, subtyping AbstractColoringAlgorithm

Constructors

AutoSparse(
    dense_ad;
    sparsity_detector=ADTypes.NoSparsityDetector(),
    coloring_algorithm=ADTypes.NoColoringAlgorithm()
)