Backends

Types

We support (and re-export) most backend choices from ADTypes.jl, and we provide a few more of our own.

Warning

Only the backends listed below are supported by DifferentiationInterface.jl, even though ADTypes.jl defines more.

Dense

ADTypes.AutoChainRules — Type

AutoChainRules{RC}

Chooses any AD library based on ChainRulesCore.jl, given an appropriate RuleConfig object.

Fields

ruleconfig::RC

ADTypes.AutoDiffractor — Type

AutoDiffractor

Chooses Diffractor.jl.

ADTypes.AutoEnzyme — Type

AutoEnzyme(Enzyme.Forward)
AutoEnzyme(Enzyme.Reverse)

Construct a forward or reverse mode AutoEnzyme backend.

AutoEnzyme{M}

Chooses Enzyme.jl.

Fields

mode::M = nothing

DifferentiationInterface.AutoFastDifferentiation — Type

AutoFastDifferentiation

Chooses FastDifferentiation.jl.

!! danger This backend is experimental, use at your own risk.

ADTypes.AutoForwardDiff — Type

AutoForwardDiff{chunksize,T}

Chooses ForwardDiff.jl.

Fields

tag::T

ADTypes.AutoForwardDiff — Method

AutoForwardDiff(; chunksize = nothing, tag = nothing)

Constructor.

ADTypes.AutoFiniteDiff — Type

AutoFiniteDiff{T1,T2,T3}

Chooses FiniteDiff.jl.

Fields

fdtype::T1 = Val(:forward)

fdjtype::T2 = fdtype

fdhtype::T3 = Val(:hcentral)

ADTypes.AutoFiniteDifferences — Type

AutoFiniteDifferences{T}

Chooses FiniteDifferences.jl.

Fields

fdm::T = nothing

ADTypes.AutoPolyesterForwardDiff — Type

AutoPolyesterForwardDiff{chunksize}

Chooses PolyesterForwardDiff.jl.

ADTypes.AutoPolyesterForwardDiff — Method

AutoPolyesterForwardDiff(; chunksize = nothing)

Constructor.

ADTypes.AutoReverseDiff — Type

AutoReverseDiff

Chooses ReverseDiff.jl.

Fields

compile::Bool = false

DifferentiationInterface.AutoTapir — Type

AutoTapir

Chooses Tapir.jl.

!! danger This backend is experimental, use at your own risk.

ADTypes.AutoTracker — Type

AutoTracker

Chooses Tracker.jl.

ADTypes.AutoZygote — Type

AutoZygote

Chooses Zygote.jl.

Sparse

DifferentiationInterface.AutoSparseFastDifferentiation — Type

AutoSparseFastDifferentiation

Chooses FastDifferentiation.jl leveraging sparsity.

!! danger This backend is experimental, use at your own risk.

ADTypes.AutoSparseFiniteDiff — Type

AutoSparseFiniteDiff

Chooses FiniteDiff.jl while exploiting sparsity.

ADTypes.AutoSparseForwardDiff — Type

AutoSparseForwardDiff{chunksize,T}

Chooses ForwardDiff.jl while exploiting sparsity.

Fields

tag::T

ADTypes.AutoSparseForwardDiff — Method

AutoSparseForwardDiff(; chunksize = nothing, tag = nothing)

Constructor.

ADTypes.AutoSparsePolyesterForwardDiff — Type

AutoSparsePolyesterForwardDiff{chunksize}

Chooses PolyesterForwardDiff.jl while exploiting sparsity.

ADTypes.AutoSparseReverseDiff — Type

AutoSparseReverseDiff

Chooses ReverseDiff.jl while exploiting sparsity.

Fields

compile::Bool = false

ADTypes.AutoSparseZygote — Type

AutoSparseZygote

Chooses Zygote.jl while exploiting sparsity.

Availability

You can use check_available to verify whether a given backend is loaded, like we did below:

backend	available

Diffractor (forward)	✅

Enzyme (forward)	✅

Enzyme (reverse)	✅

FastDifferentiation (symbolic)	✅

FiniteDiff (finite)	✅

FiniteDifferences (finite)	✅

ForwardDiff (forward)	✅

PolyesterForwardDiff (forward)	✅

ReverseDiff (reverse)	✅

Tapir (reverse)	✅

Tracker (reverse)	✅

Zygote (reverse)	✅

Mutation support

All backends are compatible with one-argument functions f(x) = y. Only some are compatible with two-argument functions f!(y, x) = nothing. You can use check_mutation to check that feature, like we did below:

backend	mutation

Diffractor (forward)	❌

Enzyme (forward)	✅

Enzyme (reverse)	✅

FastDifferentiation (symbolic)	✅

FiniteDiff (finite)	✅

FiniteDifferences (finite)	❌

ForwardDiff (forward)	✅

PolyesterForwardDiff (forward)	✅

ReverseDiff (reverse)	✅

Tapir (reverse)	✅

Tracker (reverse)	❌

Zygote (reverse)	❌

Hessian support

Only some backends are able to compute Hessians. You can use check_hessian to check that feature, like we did below:

backend	Hessian

Diffractor (forward)	❌

Enzyme (forward)	❌

Enzyme (reverse)	❌

FastDifferentiation (symbolic)	✅

FiniteDiff (finite)	✅

FiniteDifferences (finite)	✅

ForwardDiff (forward)	✅

PolyesterForwardDiff (forward)	✅

ReverseDiff (reverse)	✅

Tapir (reverse)	❌

Tracker (reverse)	❌

Zygote (reverse)	✅