Backends

Types

Most backend choices are defined by ADTypes.jl.

Warning

Only the backends listed here are supported by DifferentiationInterface.jl, even though ADTypes.jl defines more.

ADTypes.AutoChainRules — Type

AutoChainRules{RC}

Chooses any AD library based on ChainRulesCore.jl, given an appropriate RuleConfig object.

Fields

ruleconfig::RC

ADTypes.AutoDiffractor — Type

AutoDiffractor

Chooses Diffractor.jl.

ADTypes.AutoEnzyme — Type

AutoEnzyme(Enzyme.Forward)
AutoEnzyme(Enzyme.Reverse)

Construct a forward or reverse mode AutoEnzyme backend.

AutoEnzyme{M}

Chooses Enzyme.jl.

Fields

mode::M = nothing

ADTypes.AutoForwardDiff — Type

AutoForwardDiff{chunksize,T}

Chooses ForwardDiff.jl.

Fields

tag::T

ADTypes.AutoForwardDiff — Method

AutoForwardDiff(; chunksize = nothing, tag = nothing)

Constructor.

ADTypes.AutoFiniteDiff — Type

AutoFiniteDiff{T1,T2,T3}

Chooses FiniteDiff.jl.

Fields

fdtype::T1 = Val(:forward)

fdjtype::T2 = fdtype

fdhtype::T3 = Val(:hcentral)

ADTypes.AutoPolyesterForwardDiff — Type

AutoPolyesterForwardDiff{chunksize}

Chooses PolyesterForwardDiff.jl.

ADTypes.AutoPolyesterForwardDiff — Method

AutoPolyesterForwardDiff(; chunksize = nothing)

Constructor.

ADTypes.AutoReverseDiff — Type

AutoReverseDiff

Chooses ReverseDiff.jl.

Fields

compile::Bool = false

ADTypes.AutoTracker — Type

AutoTracker

Chooses Tracker.jl.

ADTypes.AutoZygote — Type

AutoZygote

Chooses Zygote.jl.

We also provide a few of our own:

DifferentiationInterface.AutoFastDifferentiation — Type

AutoFastDifferentiation

Chooses FastDifferentiation.jl.

DifferentiationInterface.SecondOrder — Type

SecondOrder

Combination of two backends for second-order differentiation.

Fields

inner::ADTypes.AbstractADType: backend for the inner differentiation

outer::ADTypes.AbstractADType: backend for the outer differentiation

Availability

You can use available to verify whether a given backend is loaded, like we did below:

Backend	available

Diffractor (forward)	✓

Enzyme (forward)	✓

Enzyme (reverse)	✓

FastDifferentiation (symbolic)	✓

FiniteDiff (forward)	✓

ForwardDiff (forward)	✓

PolyesterForwardDiff (forward)	✓

ReverseDiff (reverse)	✓

Tracker (reverse)	✓

Zygote (reverse)	✓

Mutation support

All backends are compatible with allocating functions f(x) = y. Only some are compatible with mutating functions f!(y, x) = nothing. You can use supports_mutation to check that feature, like we did below:

Backend	mutation

Diffractor (forward)	✗

Enzyme (forward)	✓

Enzyme (reverse)	✓

FastDifferentiation (symbolic)	✗

FiniteDiff (forward)	✓

ForwardDiff (forward)	✓

PolyesterForwardDiff (forward)	✓

ReverseDiff (reverse)	✓

Tracker (reverse)	✗

Zygote (reverse)	✗

Second order

For second-order differentiation, you can either

combine a pair (backend_inner, backend_outer) of inner and outer backends into a SecondOrder object

use a single backend backend, which amounts to backend_inner = backend_outer = backend

In Hessian computations, the most efficient combination is often forward-over-reverse, i.e. SecondOrder(reverse_backend, forward_backend).

Info

Many backend combinations will fail for second order. Some because of our implementation, and some because the outer backend cannot differentiate through code generated by the inner backend.

You can use supports_hessian to find working combinations, like we did below (Enzyme is skipped here due to compilation overhead):

Inner \ Outer	Diffractor (forward)	FastDifferentiation (symbolic)	FiniteDiff (forward)	ForwardDiff (forward)	PolyesterForwardDiff (forward)	ReverseDiff (reverse)	Tracker (reverse)	Zygote (reverse)

Diffractor (forward)	✗	✗	✓	✓	✓	✓	✓	✓

FastDifferentiation (symbolic)	✗	✗	✗	✗	✗	✗	✗	✗

FiniteDiff (forward)	✗	✗	✗	✓	✓	✓	✓	✓

ForwardDiff (forward)	✗	✗	✓	✓	✓	✗	✗	✗

PolyesterForwardDiff (forward)	✗	✗	✓	✓	✓	✗	✗	✗

ReverseDiff (reverse)	✗	✗	✓	✓	✓	✓	✗	✗

Tracker (reverse)	✗	✗	✓	✓	✓	✗	✗	✗

Zygote (reverse)	✗	✗	✓	✓	✓	✓	✓	✓

Package extensions

Backend-specific extension content is not part of the public API.