Interface

DifferentiationInterface.DifferentiationInterface — Module

DifferentiationInterface

An interface to various automatic differentiation backends in Julia.

Exports

AbstractBackend

AbstractForwardBackend

AbstractReverseBackend

ChainRulesForwardBackend

ChainRulesReverseBackend

EnzymeForwardBackend

EnzymeReverseBackend

FiniteDiffBackend

ForwardDiffBackend

PolyesterForwardDiffBackend

ReverseDiffBackend

ZygoteBackend

autodiff_mode

derivative

gradient

gradient!

handles_input_type

handles_output_type

handles_types

is_custom

jacobian

jacobian!

multiderivative

multiderivative!

pullback

pullback!

pushforward

pushforward!

value_and_derivative

value_and_gradient

value_and_gradient!

value_and_jacobian

value_and_jacobian!

value_and_multiderivative

value_and_multiderivative!

value_and_pullback

value_and_pullback!

value_and_pushforward

value_and_pushforward!

Utilities

Scalar to scalar

DifferentiationInterface.derivative — Method

derivative(backend, f, x) -> der

Compute the derivative der = f'(x) of a scalar-to-scalar function.

DifferentiationInterface.value_and_derivative — Function

value_and_derivative(backend, f, x) -> (y, der)

Compute the primal value y = f(x) and the derivative der = f'(x) of a scalar-to-scalar function.

Scalar to array

DifferentiationInterface.multiderivative! — Method

multiderivative!(multider, backend, f, x) -> multider

Compute the (array-valued) derivative multider = f'(x) of a scalar-to-array function, overwriting multider if possible.

DifferentiationInterface.multiderivative — Method

multiderivative(backend, f, x) -> multider

Compute the (array-valued) derivative multider = f'(x) of a scalar-to-array function.

DifferentiationInterface.value_and_multiderivative! — Function

value_and_multiderivative!(multider, backend, f, x) -> (y, multider)

Compute the primal value y = f(x) and the (array-valued) derivative multider = f'(x) of a scalar-to-array function, overwriting multider if possible.

DifferentiationInterface.value_and_multiderivative — Method

value_and_multiderivative(backend, f, x) -> (y, multider)

Compute the primal value y = f(x) and the (array-valued) derivative multider = f'(x) of a scalar-to-array function.

Array to scalar

DifferentiationInterface.gradient! — Method

gradient!(grad, backend, f, x) -> grad

Compute the gradient grad = ∇f(x) of an array-to-scalar function, overwriting grad if possible.

DifferentiationInterface.gradient — Method

gradient(backend, f, x) -> grad

Compute the gradient grad = ∇f(x) of an array-to-scalar function.

DifferentiationInterface.value_and_gradient! — Function

value_and_gradient!(grad, backend, f, x) -> (y, grad)

Compute the primal value y = f(x) and the gradient grad = ∇f(x) of an array-to-scalar function, overwriting grad if possible.

DifferentiationInterface.value_and_gradient — Method

value_and_gradient(backend, f, x) -> (y, grad)

Compute the primal value y = f(x) and the gradient grad = ∇f(x) of an array-to-scalar function.

Array to array

DifferentiationInterface.jacobian! — Method

jacobian!(jac, backend, f, x) -> jac

Compute the Jacobian matrix jac = ∂f(x) of an array-to-array function, overwriting jac if possible.

Notes

Regardless of the shape of x and y, if x has length n and y has length m, then jac is expected to be a m × n matrix. This function acts as if the input and output had been flattened with vec.

DifferentiationInterface.jacobian — Method

jacobian(backend, f, x) -> jac

Compute the Jacobian matrix jac = ∂f(x) of an array-to-array function.

Notes

Regardless of the shape of x and y, if x has length n and y has length m, then jac is expected to be a m × n matrix. This function acts as if the input and output had been flattened with vec.

DifferentiationInterface.value_and_jacobian! — Method

value_and_jacobian!(jac, backend, f, x) -> (y, jac)

Compute the primal value y = f(x) and the Jacobian matrix jac = ∂f(x) of an array-to-array function, overwriting jac if possible.

Notes

Regardless of the shape of x and y, if x has length n and y has length m, then jac is expected to be a m × n matrix. This function acts as if the input and output had been flattened with vec.

DifferentiationInterface.value_and_jacobian — Method

value_and_jacobian(backend, f, x) -> (y, jac)

Compute the primal value y = f(x) and the Jacobian matrix jac = ∂f(x) of an array-to-array function.

Notes

Regardless of the shape of x and y, if x has length n and y has length m, then jac is expected to be a m × n matrix. This function acts as if the input and output had been flattened with vec.

Primitives

Pushforward

DifferentiationInterface.pushforward! — Method

pushforward!(dy, backend::AbstractForwardBackend, f, x, dx) -> dy

Compute the Jacobian-vector product dy = ∂f(x) * dx, overwriting dy if possible.

DifferentiationInterface.pushforward — Method

pushforward(backend::AbstractForwardBackend, f, x, dx) -> dy

Compute the Jacobian-vector product dy = ∂f(x) * dx.

DifferentiationInterface.value_and_pushforward! — Method

value_and_pushforward!(dy, backend::AbstractForwardBackend, f, x, dx) -> (y, dy)

Compute the primal value y = f(x) and the Jacobian-vector product dy = ∂f(x) * dx, overwriting dy if possible.

Interface requirement

This is the only required implementation for an AbstractForwardBackend.

DifferentiationInterface.value_and_pushforward — Method

value_and_pushforward(backend::AbstractForwardBackend, f, x, dx) -> (y, dy)

Compute the primal value y = f(x) and the Jacobian-vector product dy = ∂f(x) * dx.

Pullback

DifferentiationInterface.pullback! — Method

pullback!(dx, backend::AbstractReverseBackend, f, x, dy) -> dx

Compute the vector-Jacobian product dx = ∂f(x)' * dy, overwriting dx if possible.

DifferentiationInterface.pullback — Method

pullback(backend::AbstractReverseBackend, f, x, dy) -> dx

Compute the vector-Jacobian product dx = ∂f(x)' * dy.

DifferentiationInterface.value_and_pullback! — Method

value_and_pullback!(dx, backend::AbstractReverseBackend, f, x, dy) -> (y, dx)

Compute the primal value y = f(x) and the vector-Jacobian product dx = ∂f(x)' * dy, overwriting dx if possible.

Interface requirement

This is the only required implementation for an AbstractReverseBackend.

DifferentiationInterface.value_and_pullback — Method

value_and_pullback(backend::AbstractReverseBackend, f, x, dy) -> (y, dx)

Compute the primal value y = f(x) and the vector-Jacobian product dx = ∂f(x)' * dy.

Abstract backends

DifferentiationInterface.AbstractBackend — Type

AbstractBackend

Abstract type pointing to the AD package chosen by the user, which is called a "backend".

Custom

When we say that a backend is "custom", it describes how the utilities (derivative, multiderivative, gradient and jacobian) are implemented:

Custom backends use specific routines defined in their package whenever they exist

Non-custom backends use fallbacks defined in DifferentiationInterface.jl, which end up calling the pushforward or pullback

DifferentiationInterface.AbstractForwardBackend — Type

AbstractForwardBackend <: AbstractBackend

Abstract subtype of AbstractBackend for forward mode AD packages.

DifferentiationInterface.AbstractReverseBackend — Type

AbstractReverseBackend <: AbstractBackend

Abstract subtype of AbstractBackend for reverse mode AD packages.

DifferentiationInterface.autodiff_mode — Method

autodiff_mode(backend)

Return either :forward or :reverse depending on the mode of backend.

DifferentiationInterface.handles_input_type — Method

handles_input_type(backend, ::Type{X})

Check if backend can differentiate functions with input type X.

DifferentiationInterface.handles_output_type — Method

handles_output_type(backend, ::Type{Y})

Check if backend can differentiate functions with output type Y.

DifferentiationInterface.handles_types — Method

handles_types(backend, ::Type{X}, ::Type{Y})

Check if backend can differentiate functions with input type X and output type Y.

DifferentiationInterface.is_custom — Method

is_custom(backend)

Return a boolean custom that describes how utilities (derivative, multiderivative, gradient and jacobian) are implemented.

Internals

DifferentiationInterface.basisarray — Method

basisarray(backend, a::AbstractArray, i::CartesianIndex)

Construct the i-th stardard basis array in the vector space of a with element type eltype(a).

Note

If an AD backend benefits from a more specialized basis array implementation, this function can be extended on the backend type.